### 2024-09-28 #660 @ahasverus
 forcis: An R client to access the FORCIS database
Submitting Author Name: Nicolas Casajus
Submitting Author Github Handle: @ahasverus
Other Package Authors Github handles: (comma separated, delete if none) @MatGreco90, @ChaabaneS, @xgiraud
Repository:  https://github.com/frbcesab/forcis
Version submitted: 0.1.0
Submission type: Standard

Version accepted: TBD

---

```
Package: forcis
Title: An R Client to Access the FORCIS Database
Version: 0.1.0
    person(given   = "Nicolas",
           family  = "Casajus",
           email   = "nicolas.casajus@fondationbiodiversite.fr",
    person(given   = "Mattia",
           family  = "Greco",
           role    = "aut",
           email   = "mattia_greco@outlook.com",
    person(given   = "Sonia",
           family  = "Chaabane",
           role    = "aut",
           email   = "sonia.chaabane@gmail.com",
    person(given   = "Xavier",
           family  = "Giraud",
           role    = "aut",
           email   = "giraud@cerege.fr",
    person(given   = "Thibault",
           family  = "de Garidel-Thoron",
           role    = "aut",
           email   = "garidel@cerege.fr",
    person(given   = "Khalil",
           family  = "Hammami",
           role    = "ctb",
           email   = "khalil.hammami@enetcom.usf.tn"))
Description: Provides an interface to the FORCIS database
    (<https://zenodo.org/doi/10.5281/zenodo.7390791>) on global foraminifera
    distribution. This package allows to download and to handle FORCIS data.
    It is part of the FRB-CESAB working group FORCIS.
    <https://www.fondationbiodiversite.fr/en/the-frb-in-action/programs-and-projects/le-cesab/forcis/>.
    sf,
    vroom
    fs,
    withr
```

> This package is designed to download the FORCIS data files hosted on [Zenodo](https://zenodo.org/doi/10.5281/zenodo.7390791). It includes functions to download (_data retrieval_), select, filter, reshape, and visualize data (_data munging_).

> This package should be of interest to scientists working on Foraminifera species distribution and interested in the FORCIS database (spatial analyses, time series analyses, etc.). The package have been developed to facilitate the data wrangling to avoid some pitfalls and to easily get data ready to be analyzed/visualized.

> No other package exists to handle the FORCIS database. Note that we are authors of the database and already published a [Data paper](https://www.nature.com/articles/s41597-023-02264-2) describing the database.

> Not applicable.

> Pre-submission inquiry: [#655](https://github.com/ropensci/software-review/issues/655)
> Editor: @adamhsparks

> The function `pkgcheck::pkgcheck()` returns the following report:
>
> ```
> ── forcis 0.1.0 ────────────────────────────────────────────
>
> ✔ Package name is available
> ✔ has a 'codemeta.json' file.
> ✔ has a 'contributing' file.
> ✔ uses 'roxygen2'.
> ✔ 'DESCRIPTION' has a URL field.
> ✔ 'DESCRIPTION' has a BugReports field.
> ✔ Package has at least one HTML vignette
> ✔ All functions have examples.
> ✔ Package has continuous integration checks.
> ✔ Package coverage is 97.4%.
> ✔ R CMD check found no errors.
> ✔ R CMD check found no warnings.
>
> ℹ Current status:
> ✔ This package may be submitted.
> ```
>
> The package `goodpractice` returns warnings:
>   - **Write unit tests**: some functions are difficult to test (HTTP requests)
>   - **Avoid calling `setwd()`**: this function is used in unit tests in combination with `withr::defer()`
