+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   forcis: forcis.html
   sampling-devices: sampling-devices.html
   select-and-filter-data: select-and-filter-data.html
-last_built: 2024-09-28T09:25Z
+last_built: 2024-09-28T09:39Z
 urls:
   reference: https://frbcesab.github.io/forcis/reference
   article: https://frbcesab.github.io/forcis/articles
