### 2024-10-09 #660 @beatrizmilz
- [x] **Project management**: Are the issue and PR trackers in a good shape, e.g. are there outstanding bugs, is it clear when feature requests are meant to be tackled? **(there are no open issues)**
---

Hi @ahasverus ! Congratulations for you and the team for this great package, and also for the [publication of the data paper on Nature] (https://www.nature.com/articles/s41597-023-02264-2).

The comments below are related to the `Editor Checks` above.

- (1) In the vignette ["Database versions"](https://frbcesab.github.io/forcis/articles/database-versions.html), there is a paragraph that is not clear to me:
> A good practice to ensure that the user works with the latest version of the database might be to add this line at the beginning of the script:
```
download_forcis_db(version = NULL, ...)
```

---
- (2) In the vignette [Select, reshape, and filter data](https://frbcesab.github.io/forcis/articles/select-and-filter-data.html): it's not clear to me what the required columns mean. I saw the page for the function [`get_required_columns()`](https://frbcesab.github.io/forcis/reference/get_required_columns.html), and there is a list of columns. But could it be possible to describe a bit more? Why are they required? Is that because these columns are the most important for basic analysis?

---

- (3) In the vignettes (for example [Select, reshape, and filter data](https://frbcesab.github.io/forcis/articles/select-and-filter-data.html)): The names of the sections are the names of functions. I think it's best to name sections with a description of the task.
You can see some examples in vignettes of other packages in rOpenSci: [magick](https://docs.ropensci.org/magick/articles/intro.html), [tabulapdf](https://docs.ropensci.org/tabulapdf/articles/tabulapdf.html), etc.

How it is now:
```
On this page

    Setup
    select_taxonomy()
    select_forcis_columns()
    filter_by_month()
    filter_by_year()
    filter_by_bbox()
    filter_by_ocean()
    filter_by_polygon()
    filter_by_species()
    convert_to_long_format()
```
Idea: (this is just an example)
```

    Setup
    Selecting columns
       Selecting columns by taxonomy
       Selecting required columns
    Filtering rows
       Filter by month of data collection
       Filter by year of data collection
       Filter by location (bounding box)
       Filter by ocean
       Filter by polygon
       Filter by species
    Reshaping
       Convert to long format
```
---
- (4) About tests for fuctions that creates plots, in the [dev guide](https://devguide.ropensci.org/pkg_building.html#testing) is said:
> For testing your functions creating plots, we suggest using [vdiffr](https://vdiffr.r-lib.org/), an extension of the testthat package that relies on [testthat snapshot tests](https://testthat.r-lib.org/articles/snapshotting.html).

From what I checked in some tests (eg. [test-plot_record_by_month.R](https://github.com/FRBCesab/forcis/blob/master/tests/testthat/test-plot_record_by_month.R)), the tests verify the class of the plot created.
Could you improve the tests for the functions that create plots using  [vdiffr](https://vdiffr.r-lib.org/)?

---

- (5) About testing the download functions: I looked up the tests for the download function but I'm not sure if it follows the best practices. I need some time to read more about testing functions that access resources on the web (eg. the [dev guide](https://devguide.ropensci.org/pkg_building.html#testing) recommends the book [HTTP testing in R](https://books.ropensci.org/http-testing/index.html)). As you wrote in the submission, "some functions are difficult to test (HTTP requests)". I'll be back with that feedback soon!
---

- (6) This is more of a question. In the package webpage, it says that the package has been developed for the **Centre for the Synthesis and Analysis of Biodiversity**. Does this mean that they are the funder? If so, they can be added as `fnd` in the authors list. [This post](https://ropensci.org/blog/2018/03/16/thanking-reviewers-in-metadata/) can be useful to understand these three-letter-code used in the authors list on DESCRIPTION.
