### 2024-10-10 #660 @beatrizmilz
> Hi @beatrizmilz!
>
>
>
> Thank you for this first round of comments. I will start looking into them in the next few days and come back to you as soon as possible.

Hi! I hope the comments are helpful.
Most of the comments are suggestions, feel free to work on the ones that makes sense for you.

The comments about testing are the most important, since they are recommendations from the dev guide!
