+++ NEWS.md
@@ -1,3 +1,8 @@
+# forcis 0.1.0.9000
+
+Answer to editor and reviewer comments.
+
+
 # forcis 0.1.0
 
 First stable release.
