+++ README.Rmd
@@ -106,7 +106,7 @@ For an overview of the main features of `forcis`, please read the [Get started](
 
 Please cite this package as: 
 
-> Casajus N, Greco M, Chaabane S, Giraud X & de Garidel-Thoron T (2024) forcis: An R client to access the FORCIS database. R package version 0.1.0. URL: https://frbcesab.github.io/forcis/.
+> Casajus N, Greco M, Chaabane S, Giraud X & de Garidel-Thoron T (2024) forcis: An R client to access the FORCIS database. R package version 0.1.0.9000. URL: https://frbcesab.github.io/forcis/.
 
 You can also run:
 
