+++ README.md
@@ -108,7 +108,7 @@ Please cite this package as:
 
 > Casajus N, Greco M, Chaabane S, Giraud X & de Garidel-Thoron T (2024)
 > forcis: An R client to access the FORCIS database. R package version
-> 0.1.0. URL: <https://frbcesab.github.io/forcis/>.
+> 0.1.0.9000. URL: <https://frbcesab.github.io/forcis/>.
 
 You can also run:
 
