+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: forcis
 Type: Package
 Title: An R Client to Access the FORCIS Database
-Version: 0.1.0
+Version: 0.1.0.9000
 Authors@R: c(
     person(given   = "Nicolas",
            family  = "Casajus",
