+++ NEWS.md
@@ -1,6 +1,11 @@
 # forcis 0.1.0.9000
 
-Answer to editor and reviewer comments.
+Answers to editor and reviewer comments.
+
+* Improve documentation (#64, #65 & #66)
+* Add funder in `DESCRIPTION` (#69)
+* Improve unit tests (#67 & #68)
+
 
 
 # forcis 0.1.0
