### 2024-10-14 b0c9402
doc: edit changelog (add urls)
