### 2024-10-14 #660 @ahasverus
Hi @beatrizmilz!

No, all your comments are very useful.
I have started to work on some of them.

<hr>

> - (1) In the vignette ["Database versions"](https://frbcesab.github.io/forcis/articles/database-versions.html), there is a paragraph that is not clear to me:
>
>> A good practice to ensure that the user works with the latest version of the database might be to add this line at the beginning of the script:
>
> ```r
> download_forcis_db(version = NULL, ...)
> ```

**Answer:** Thanks for reporting the lack of clarity of this section. I have added a few sentences to clarify this paragraph in the vignette [Database versions](https://frbcesab.github.io/forcis/articles/database-versions.html) (commit [3ab5baa](https://github.com/FRBCesab/forcis/commit/3ab5baa2902da11cf3ef6fab913d4e91d16f039c)). I hope it's clearer.

<hr>

> - (2) In the vignette [Select, reshape, and filter data](https://frbcesab.github.io/forcis/articles/select-and-filter-data.html): it's not clear to me what the required columns mean. I saw the page for the function [get_required_columns()](https://frbcesab.github.io/forcis/reference/get_required_columns.html), and there is a list of columns. But could it be possible to describe a bit more? Why are they required? Is that because these columns are the most important for basic analysis?

**Answer:** I have added a few sentences to explain why these columns are required in the vignette [Select, reshape, and filter data](https://frbcesab.github.io/forcis/articles/select-and-filter-data.html) and in the documentation of the function [`get_required_columns()`](https://frbcesab.github.io/forcis/reference/get_required_columns.html) (commit [06a053b](https://github.com/FRBCesab/forcis/commit/06a053bdf7fdb01873551991f2a01d4728474943)).

<hr>

> - (3) In the vignettes (for example [Select, reshape, and filter data](https://frbcesab.github.io/forcis/articles/select-and-filter-data.html)): The names of the sections are the names of functions. I think it's best to name sections with a description of the task.
You can see some examples in vignettes of other packages in rOpenSci: [magick](https://docs.ropensci.org/magick/articles/intro.html), [tabulapdf](https://docs.ropensci.org/tabulapdf/articles/tabulapdf.html), etc.

**Answer:** Thanks for this suggestion. I have modified the section names in the vignettes [Select, reshape, and filter data](https://frbcesab.github.io/forcis/articles/select-and-filter-data.html) and [Data visualization](https://frbcesab.github.io/forcis/articles/data-visualization.html) (commit [81db6ae](https://github.com/FRBCesab/forcis/commit/81db6ae33aa6e6bf2d5fb928e5e959bca12c55d7)).

<hr>

> - (6) This is more of a question. In the package webpage, it says that the package has been developed for the Centre for the Synthesis and Analysis of Biodiversity. Does this mean that they are the funder? If so, they can be added as fnd in the authors list. [This post](https://ropensci.org/blog/2018/03/16/thanking-reviewers-in-metadata/) can be useful to understand these three-letter-code used in the authors list on `DESCRIPTION`.

**Answer:** Indeed, the research group [FORCIS](https://www.fondationbiodiversite.fr/en/the-frb-in-action/programs-and-projects/le-cesab/forcis/) has been funded by the FRB-CESAB. I have added it to the `DESCRIPTION` file with the role `fnd` (commit [6bd6e2b](https://github.com/FRBCesab/forcis/commit/6bd6e2be92a8e1b4b1d6287e6802e3f4d1c283d1)).

<hr>

Regarding your comments (4) and (5), I need to read more about the packages `vdiffr` and `httptest` to improve and implement unit tests for plotting functions and HTTP requests.

I will come back to you very soon.
Thanks again.
