+++ tests/testthat/test-plot_record_by_month.R
@@ -26,5 +26,7 @@ test_that("Test plot_record_by_month() for success", {
   
   expect_true("gg" %in% class(gg))
   expect_true("ggplot" %in% class(gg))
+  
+  vdiffr::expect_doppelganger("Plot record by month", gg)
 })
 
