+++ tests/testthat/test-plot_record_by_season.R
@@ -28,4 +28,6 @@ test_that("Test plot_record_by_season() for success", {
   
   expect_true("gg" %in% class(gg))
   expect_true("ggplot" %in% class(gg))
+  
+  vdiffr::expect_doppelganger("Plot record by season", gg)
 })
