+++ CITATION.cff
@@ -34,6 +34,18 @@ authors:
   name-particle: de
   email: garidel@cerege.fr
   orcid: https://orcid.org/0000-0001-8983-9571
+preferred-citation:
+  type: manual
+  title: 'forcis: An R client to access the FORCIS database'
+  authors:
+  - name: Casajus Nicolas
+  - name: Greco Mattia
+  - name: Chaabane Sonia
+  - name: Giraud Xavier
+  - name: de Garidel-Thoron Thibault
+  year: '2024'
+  notes: R package version 0.1.0.9000
+  url: https://frbcesab.github.io/forcis
 repository-code: https://github.com/FRBCesab/forcis
 url: https://frbcesab.github.io/forcis
 contact:
@@ -42,6 +54,62 @@ contact:
   email: nicolas.casajus@fondationbiodiversite.fr
   orcid: https://orcid.org/0000-0002-5537-5294
 references:
+- type: article
+  title: 'The FORCIS database: A global census of planktonic Foraminifera from ocean
+    waters'
+  authors:
+  - name: Chaabane Sonia
+  - name: De Garidel-Thoron Thibault
+  - name: Giraud Xavier
+  - name: Schiebel Ralf
+  - name: Beaugrand Gregory
+  - name: Brummer Geert-Jan
+  - name: Casajus Nicolas
+  - name: Greco Mattia
+  - name: Grigoratou Maria
+  - name: Howa Hélène
+  - name: Jonkers Lukas
+  - name: Kucera Michal
+  - name: Kuroyanagi Azumi
+  - name: Meilland Julie
+  - name: Monteiro Fanny
+  - name: Mortyn Graham
+  - name: Almogi-Labin Ahuva
+  - name: Asahi Hirofumi
+  - name: Avnaim-Katav Simona
+  - name: Bassinot Franck
+  - name: Davis Catherine V.
+  - name: Field David B.
+  - name: Hernández-Almeida Iván
+  - name: Herut Barak
+  - name: Hosie Graham
+  - name: Howard Will
+  - name: Jentzen Anna
+  - name: Johns David G.
+  - name: Keigwin Lloyd
+  - name: Kitchener John
+  - name: Kohfeld Karen E.
+  - name: Lessa Douglas V. O.
+  - name: Manno Clara
+  - name: Marchant Margarita
+  - name: Ofstad Siri
+  - name: Ortiz Joseph D.
+  - name: Post Alexandra
+  - name: Rigual-Hernandez Andres
+  - name: Rillo Marina C.
+  - name: Robinson Karen
+  - name: Sagawa Takuya
+  - name: Sierro Francisco
+  - name: Takahashi Kunio T.
+  - name: Torfstein Adi
+  - name: Venancio Igor
+  - name: Yamasaki Makoto
+  - name: Ziveri Patrizia
+  journal: Scientific Data
+  volume: '10'
+  doi: 10.1038/s41597-023-02264-2
+  year: '2023'
+  start: '354'
 - type: software
   title: 'R: A Language and Environment for Statistical Computing'
   notes: Depends
@@ -293,6 +361,31 @@ references:
   year: '2024'
   doi: 10.32614/CRAN.package.testthat
   version: '>= 3.0.0'
+- type: software
+  title: vdiffr
+  abstract: 'vdiffr: Visual Regression Testing and Graphical Diffing'
+  notes: Suggests
+  url: https://vdiffr.r-lib.org/
+  repository: https://CRAN.R-project.org/package=vdiffr
+  authors:
+  - family-names: Henry
+    given-names: Lionel
+    email: lionel@posit.co
+  - family-names: Pedersen
+    given-names: Thomas Lin
+    email: thomas.pedersen@posit.co
[…]
