+++ DESCRIPTION
@@ -62,5 +62,6 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
+    vdiffr,
     withr
 Config/testthat/edition: 3
