### 2024-10-20 24142ce
Update CITATION.cff
