### 2025-02-02 #660 @maurolepore
Dear @ahasverus FYI I just started my EiC rotation. I see you responded to many of @beatrizmilz's comments. Is there anything else you would like to do before @beatrizmilz considers your changes and moves on to seeking reviewers?
