+++ R/get_available_versions.R
@@ -24,14 +24,16 @@ get_available_versions <- function() {
   ## Retrieve information ----
   
   meta <- get_metadata()
-  meta <- meta$"hits"$"hits"$"metadata"
+  meta <- lapply(meta$"hits"$"hits", function(x) {
+    data.frame("publication_date" = x$"metadata"$"publication_date",
+               "version"          = x$"metadata"$"version",
+               "access_right"     = x$"metadata"$"access_right")
+  })
+  
+  meta <- do.call(rbind.data.frame, meta)
   
   
   ## Clean output ----
-  
-  meta <- data.frame("publication_date" = meta$"publication_date",
-                     "version"          = meta$"version",
-                     "access_right"     = meta$"access_right")
   
   meta <- meta[order(as.Date(meta$"publication_date"), decreasing = TRUE), ]
   
