### 2025-02-04 586ecc7
ci: update test coverage action
