### 2025-02-04 053a879
ci: update test coverage action
