### 2025-03-05 #660 @beatrizmilz
> Hi [@maurolepore](https://github.com/maurolepore)
>
> I have just adapted unit tests for HTTP requests as suggested in this comment:
>
> > (5) About testing the download functions: I looked up the tests for the download function but I'm not sure if it follows the best practices. I need some time to read more about testing functions that access resources on the web (eg. the [dev guide](https://devguide.ropensci.org/pkg_building.html#testing) recommends the book [HTTP testing in R](https://books.ropensci.org/http-testing/index.html)). As you wrote in the submission, "some functions are difficult to test (HTTP requests)". I'll be back with that feedback soon!
>
> **Answer:** Done in commit [15383d8](https://github.com/FRBCesab/forcis/commit/15383d8f20407e9d7bad5a4431c2768a92d1bfeb) using the [httptest2](https://enpiar.com/httptest2/) package.
>
> I think we've answered all the comments raised by [@beatrizmilz](https://github.com/beatrizmilz) and we'are ready for the next steps of the review.
>
> Thanks! Nicolas

Great! @ahasverus
Sorry for the delay, Nicolas.
And thank you for improvements! I will move forward with the review process.
