### 2025-03-05 92e7d1b
doc: add ropensci badge [no ci]
