### 2025-03-06 #660 @beatrizmilz
> [@beatrizmilz](https://github.com/beatrizmilz) [@ahasverus](https://github.com/ahasverus) I don't know why the `check package` results didn't appear there. The logs seems to indicate a glitch with the GitHub cli. I'll re-trigger them manually ...

Thank you Mark!
