### 2025-03-06 #660 @mpadge
(Note also that the check for default "master" branch has been implemented since the previous checks were issues; this should be addressed. We're working on updating [our Dev Guide to describe that](https://github.com/ropensci/dev_guide/pull/863), but in the meantime, [this blog post](https://www.tidyverse.org/blog/2021/10/renaming-default-branch/) might help.)
