### 2025-03-06 #660 @ahasverus
Hi @beatrizmilz & @mpadge !

I've renamed the default branch `master` to [`main`](https://github.com/FRBCesab/forcis/tree/main).

Regarding the comment on the lint message, I confirm that the call to `library()` only appears in function documentations (example sections), vignettes, README, and tests.
