### 2025-03-19 #660 @sckott
- AFAICT the guidelines are followed

Estimated hours spent reviewing: 1.5 hrs

  -  No thanks

---

Nice work on this package!

Tests:
- All passed except for `test-download_forcis_db.R` which had error:
```
Error in `stop_request(req)`: An unexpected request was made:
GET https://zenodo.org/api/records/?q=conceptrecid%3A7390791&all_versions=true
Expected mock file: zenodo.org/api/records-f5a5db.*
```

In running tests I get quite a lot of warnings about tidyselect like

```r
Warning (test-compute_abundances.R:25:3): Test compute_abundances() for success
Use of .data in tidyselect expressions was deprecated in tidyselect 1.2.0.
i Please use `"to_drop"` instead of `.data$to_drop`
```

Other comments:
- **A statement of need**: README does not address who the target audience is
- I have a feeling CRAN maintainers will complain about this hidden dot file `.forcis` described here https://frbcesab.github.io/forcis/articles/database-versions.html#using-a-specific-version - They generally do not like it when any files are written to the users machine unless they're in a temporary directory. Though you may get away with it if during anything that's run during r cmd check (examples, tests, etc.) you write that file to a temp dir.
- for options used by this pkg, i'd probably want to add a prefix to each of them so they don't conflict with any other potential options the user may have in their session, eg., `check_for_update` could be `forcis_check_for_update`. the other option I see you use already has that prefix
- there are a number of checks for required parameters in functions even though you have no default for those ame parameters - you shouldn't need this check inside the function in these cases
- I wonder if it's intended that there's a mix of usage of returning data.frame's and tibbles. Given the large numbers of columns and rows - and that you're already using tidyverse pkgs here - it might be worth always returning tibbles so users can get a quick summary of the data and not have the console covered by a long list of columns. Just a thought.
- I see that you have a "style guide" section in your Contributing.md file. That's great, but I wonder if it might make your life easier to use automated styling/formatting if you're not doing so already (via `styler` or [`air`](https://github.com/posit-dev/air)). I can see you have many contributors; automated  styling/formatting will make it easier to works on a single codebase across many contributors.
- In `get_current_version` you do `readLines(".forcis")` to check the database version. However, what if the user doesn't have the data anymore? That is, the data got deleted somehow but they still have the .forcis file. I imagine you'd want to check to make sure they still have the data before reporting the version? Or maybe not?
- For where to store data files have you considered https://rappdirs.r-lib.org/ ?
- Just a thought, not a required change: There's quite a bit of duplicated code in the `read_*` functions. I would suggest somehow reducing the amount of lines of duplicated code. One approach might be to make little helper functions for the code that's used across each function. Another approach would be to encapsulate the lines that are mostly in common into a separate function and use that, e.g., the below change passes your tests, and then you could use `read_data_` with the other `read_*` functions too.

<details><summary>click to expand</summary>

```r
read_cpr_north_data <- function(
  path = ".",
  version = options()$"forcis_version",
  check_for_update = options()$"check_for_update"
) {
  path <- read_data_(
    check_file_fun = cpr_north_filename,
    data_msg = "North CPR",
    path = path,
    version = version,
    check_for_update = check_for_update
  )

  ## Read data ----
  file_name <- list.files(path, pattern = cpr_north_filename())
  data <- vroom::vroom(
    file.path(path, file_name),
    delim = ";",
    altrep = FALSE,
    show_col_types = FALSE
  )
  data <- as.data.frame(data)
  data <- add_data_type(data, "CPR North")

  ## Check and convert columns ----
  taxa_columns <- c("count_bin_min", "count_bin_max")
  for (i in seq_len(length(taxa_columns))) {
    check_field_in_data(data, taxa_columns[i])
    data[, taxa_columns[i]] <- as.numeric(data[, taxa_columns[i]])
  }

  data
}

read_data_ <- function(
  check_file_fun,
  data_msg,
  path = ".",
  version = options()$"forcis_version",
  check_for_update = options()$"check_for_update"
) {
  ## Check args ----
  check_if_character(path)
  check_version(version)

  ## Check/set version ----
  version <- set_version(version, ask = FALSE)

  ## Check local database ----
  path <- file.path(path, "forcis-db", paste0("version-", version))

  if (!dir.exists(path)) {
    stop(
      "The directory '",
      path,
      "' does not exist. Please check the ",
      "argument 'path' or use the function 'download_forcis_db()'.",
      call. = FALSE
    )
  }

  ## Check file ----
  file_name <- list.files(path, pattern = check_file_fun())
  if (!length(file_name)) {
    stop(
      sprintf(
        "The %s dataset does not exist. Please use the function ",
        data_msg
      ),
      "'download_forcis_db()'.",
      call. = FALSE
    )
  }

  ## Check for update ----
  if (is.null(check_for_update)) {
    check_for_update <- TRUE
  }

  if (check_for_update) {
    if (version != get_latest_version()) {
      message(
        "A newer version of the FORCIS database is available. Use ",
        "'download_forcis_db(version = NULL)' to download it."
      )
    }
  }

  path
}
```
