### 2025-03-20 FRBCesab/forcis#70 @ahasverus
rOpenSci review: Unit tests failed
As mentionned in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049), some unit tests in `test-download_forcis_db.R` fail and return the following error message:

```
Error in `stop_request(req)`: An unexpected request was made:
GET https://zenodo.org/api/records/?q=conceptrecid%3A7390791&all_versions=true
Expected mock file: zenodo.org/api/records-f5a5db.*
```

Need to fix this.
