### 2025-03-20 FRBCesab/forcis#71 @ahasverus
rOpenSci review: Warnings in unit tests
As mentionned in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049), some unit tests return warnings liked to the `tidyselect` package. Especially:

```
Warning (test-compute_abundances.R:25:3): Test compute_abundances() for success
Use of .data in tidyselect expressions was deprecated in tidyselect 1.2.0.
i Please use `"to_drop"` instead of `.data$to_drop`
```
