### 2025-03-20 FRBCesab/forcis#72 @ahasverus
rOpenSci review: Add a statement of need
As mentionned in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049), the `README` should contain a **Statement of need** explaining the audience targetted by the package.
