### 2025-03-20 FRBCesab/forcis#73 @ahasverus
rOpenSci review: Hidden dotfile .forcis
The following statement has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049):

> I have a feeling CRAN maintainers will complain about this hidden dotfile `.forcis` described [here](https://frbcesab.github.io/forcis/articles/database-versions.html#using-a-specific-version). They generally do not like it when any files are written to the users machine unless they're in a temporary directory. Though you may get away with it if during anything that's run during r cmd check (examples, tests, etc.) you write that file to a temp dir.

Maybe we should consider not using this way of handling database version.
