### 2025-03-20 FRBCesab/forcis#74 @ahasverus
rOpenSci review: Add prefix to global options
As mentionned in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049), global options created and used by the package should always have a prefix (e.g. `forcis_*`)

> For options used by this pkg, i'd probably want to add a prefix to each of them so they don't conflict with any other potential options the user may have in their session, eg., check_for_update could be forcis_check_for_update. the other option I see you use already has that prefix
