### 2025-03-20 FRBCesab/forcis#75 @ahasverus
rOpenSci review: Functions should always return tibbles
As mentionned in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049), there is a mix of `data.frame` and `tibble` in function returns.

> I wonder if it's intended that there's a mix of usage of returning data.frame's and tibbles. Given the large numbers of columns and rows - and that you're already using tidyverse pkgs here - it might be worth always returning tibbles so users can get a quick summary of the data and not have the console covered by a long list of columns. Just a thought.

All functions should return `tibble`.
