### 2025-03-20 FRBCesab/forcis#76 @ahasverus
rOpenSci review: Coding style
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049):

> I see that you have a "style guide" section in your `Contributing.md` file. That's great, but I wonder if it might make your life easier to use automated styling/formatting if you're not doing so already (via `styler` or [air](https://github.com/posit-dev/air)). I can see you have many contributors; automated styling/formatting will make it easier to works on a single codebase across many contributors.

We will use [air](https://github.com/posit-dev/air) on Positron IDE and change the "Style guide" section in the `Contributing.md` file to encourage future contributors to use the same formatter.
