### 2025-03-20 FRBCesab/forcis#77 @ahasverus
rOpenSci review: .forcis present but data deleted
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049):

> In `get_current_version()` you do `readLines(".forcis")` to check the database version. However, what if the user doesn't have the data anymore? That is, the data got deleted somehow but they still have the `.forcis` file. I imagine you'd want to check to make sure they still have the data before reporting the version? Or maybe not?

Maybe outdated if #73 is solved by removing the usage of the `.forcis` anymore.
