### 2025-03-20 FRBCesab/forcis#78 @ahasverus
rOpenSci review: Code factoring
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2737661049):

> Just a thought, not a required change: There's quite a bit of duplicated code in the `read_*()` functions. I would suggest somehow reducing the amount of lines of duplicated code. One approach might be to make little helper functions for the code that's used across each function. Another approach would be to encapsulate the lines that are mostly in common into a separate function and use that, e.g., the below change passes your tests, and then you could use `read_data_()` with the other read_* functions too.

Suggestion made by the reviewer:

```r
read_cpr_north_data <- function(
  path = ".",
  version = options()$"forcis_version",
  check_for_update = options()$"check_for_update"
) {
  path <- read_data_(
    check_file_fun = cpr_north_filename,
    data_msg = "North CPR",
    path = path,
    version = version,
    check_for_update = check_for_update
  )

  ## Read data ----
  file_name <- list.files(path, pattern = cpr_north_filename())
  data <- vroom::vroom(
    file.path(path, file_name),
    delim = ";",
    altrep = FALSE,
    show_col_types = FALSE
  )
  data <- as.data.frame(data)
  data <- add_data_type(data, "CPR North")

  ## Check and convert columns ----
  taxa_columns <- c("count_bin_min", "count_bin_max")
  for (i in seq_len(length(taxa_columns))) {
    check_field_in_data(data, taxa_columns[i])
    data[, taxa_columns[i]] <- as.numeric(data[, taxa_columns[i]])
  }

  data
}

read_data_ <- function(
  check_file_fun,
  data_msg,
  path = ".",
  version = options()$"forcis_version",
  check_for_update = options()$"check_for_update"
) {
  ## Check args ----
  check_if_character(path)
  check_version(version)

  ## Check/set version ----
  version <- set_version(version, ask = FALSE)

  ## Check local database ----
  path <- file.path(path, "forcis-db", paste0("version-", version))

  if (!dir.exists(path)) {
    stop(
      "The directory '",
      path,
      "' does not exist. Please check the ",
      "argument 'path' or use the function 'download_forcis_db()'.",
      call. = FALSE
    )
  }

  ## Check file ----
  file_name <- list.files(path, pattern = check_file_fun())
  if (!length(file_name)) {
    stop(
      sprintf(
        "The %s dataset does not exist. Please use the function ",
        data_msg
      ),
      "'download_forcis_db()'.",
      call. = FALSE
    )
  }

  ## Check for update ----
  if (is.null(check_for_update)) {
    check_for_update <- TRUE
  }

  if (check_for_update) {
    if (version != get_latest_version()) {
      message(
        "A newer version of the FORCIS database is available. Use ",
        "'download_forcis_db(version = NULL)' to download it."
      )
    }
  }

  path
}
```
