+++ R/compute_concentrations.R
@@ -33,9 +33,6 @@
 #' 
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #' 
-#' # Add 'data_type' column ----
-#' net_data$"data_type" <- "Net"
-#' 
 #' # Select a taxonomy ----
 #' net_data <- select_taxonomy(net_data, taxonomy = "VT")
 #' 
@@ -67,14 +64,17 @@ compute_concentrations <- function(data, aggregate = TRUE) {
   
   if (get_data_type(data) == "CPR North") {
       
-    return(
-      data %>% 
-        mutate(min_conc_binned = .data$count_bin_min / 
-                 .data$sample_volume_filtered,
-               max_conc_binned = .data$count_bin_max / 
-                 .data$sample_volume_filtered) %>% 
-        select(-c(.data$count_bin_min,
-                  .data$count_bin_max)))
+    data[["min_conc_binned"]] <- 
+      data$"count_bin_min" / data$"sample_volume_filtered"
+
+    data[["max_conc_binned"]] <- 
+      data$"count_bin_max" / data$"sample_volume_filtered"
+
+    cols_to_remove <- c("count_bin_min", "count_bin_max")
+  
+    data <- data[ , !(colnames(data) %in% cols_to_remove)]
+  
+    return(data)
   }
   
   check_unique_taxonomy(data)
@@ -82,107 +82,169 @@ compute_concentrations <- function(data, aggregate = TRUE) {
   
   taxa_cols <- get_species_names(data) 
   
-  ready_dat <- data %>%
-    filter(.data$subsample_count_type == "Absolute") %>%  
-    pivot_longer(all_of(taxa_cols), 
-                 names_to  = 'taxa', 
-                 values_to = 'counts') %>%
-    mutate(to_drop = ifelse(is.na(.data$counts), 'drop', 'keep')) %>% 
-    filter(.data$to_drop == 'keep') %>%
-    select(-c(.data$to_drop,
-              .data$subsample_count_type,
-              .data$subsample_all_shells_present_were_counted,
-              .data$total_of_forams_counted_ind,
-              .data$sampling_device_type))%>% 
-    rename('counts_n_conc' = 'counts') 
   
-  abs_data_to_convert <- data %>%
-    filter(.data$sample_volume_filtered > 0) %>% 
-    filter(.data$subsample_count_type == 'Raw') %>%   
-    pivot_longer(all_of(taxa_cols), 
-                 names_to  = 'taxa', 
-                 values_to = 'counts') %>%
-    mutate(to_drop = ifelse(is.na(.data$counts), 'drop', 'keep')) %>% 
-    filter(.data$to_drop == 'keep') %>% 
-    select(-.data$to_drop) %>% 
-    mutate(new_counts = .data$counts / .data$sample_volume_filtered) %>% 
-    select(-c(.data$counts,
-              .data$subsample_count_type,
-              .data$subsample_all_shells_present_were_counted,
-              .data$total_of_forams_counted_ind,
-              .data$sampling_device_type)) %>% 
-    rename('counts_n_conc' = 'new_counts') %>% 
-    distinct()
+  ## Absolute data ----
   
-  rel_data_to_convert <- data %>%
-    filter(.data$sample_volume_filtered > 0) %>%
-    filter(.data$subsample_count_type == 'Relative') %>%
-    pivot_longer(all_of(taxa_cols), 
-                 names_to  = 'taxa', 
-                 values_to = 'counts') %>% 
-    mutate(to_drop = ifelse(is.na(.data$counts), 'drop', 'keep')) %>%
[…]
