### 2025-03-27 9bc69e1
factor: remove dependency to dplyr
