### 2025-03-27 550216c
build: remove dplyr & add tibble dependency
