+++ DESCRIPTION
@@ -49,11 +49,11 @@ VignetteBuilder: knitr
 Depends: 
     R (>= 2.10)
 Imports: 
-    dplyr,
     ggplot2,
     httr2,
     rlang,
     sf,
+    tibble,
     tidyr,
     utils,
     vroom
