### 2025-03-27 7eae3b4
doc: update function examples
