+++ man/computations.Rd
@@ -48,9 +48,6 @@ file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 
 net_data <- read.table(file_name, dec = ".", sep = ";")
 
-# Add 'data_type' column ----
-net_data$"data_type" <- "Net"
-
 # Select a taxonomy ----
 net_data <- select_taxonomy(net_data, taxonomy = "VT")
 
