+++ man/plot_record_by_season.Rd
@@ -25,9 +25,6 @@ file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 
 net_data <- read.table(file_name, dec = ".", sep = ";")
 
-# Add 'data_type' column ----
-net_data$"data_type" <- "Net"
-
 # Plot data by year (example dataset) ----
 plot_record_by_season(net_data)
 }

[…]
