### 2025-03-27 9926fdc
Update CITATION.cff
