### 2025-03-27 #660 @heyairf
---

Great work on this package! I love the logo.

I started my review by going through the documentation with very little knowledge of the field, so my perspective may not represent that of the intended user base. I hope my feedback is helpful. I would also like to self identify as a non-expert in R package development.. therefore, my review primarily focuses on the documentation/user experience side.

- README includes an overview of the what it does but does not indicate why it was made or who it is for. A sentence or two on how this data could be used might be a nice addition to overview section, though once the target audience is identified this might not be necessary.

- I think a small example of the package functionality on the README page or the Get Started page would be a great addition. I had to click around a lot before I understood what the package use would look like, even after running the code on the get started page.

##### Get Started page

- I was still pretty unsure what the purpose and final result of using this package was at the end of this page. I recommend including a small example here! Even as simple as a `head()` and small description of what I am looking at would provide a bit more context to users.

- Linking to a short vignette describing the taxonomy codes (similar in scope to the Sampling devices vignette) might be less intimidating than linking to the data paper... or at least providing and indication of which section in the data paper to read about this would be helpful! It was not immediately clear from the paper headings where to find this information.

##### Select, reshape, filter

- In the selecting columns section, it was unclear to me if you were showing two different selection methods, or a multistep selection method... are you showing a filter of VT data and then forcis columns? perhaps naming the objects differently and/or expanding the commented lines would make this clearer.
- Also, the same selection code was used on the Get Started page but the object was given a different name... could consider keeping this consistent.

##### Conversion / visualization

- These vignettes were the first time I felt I understood the point of the package/some potential uses. Having a small visualization example like these on the README or Get Started to show a start to finish workflow to a sensible end result would be helpful even if its redundant with the more detailed vignettes.

##### Function documentation and examples

- looks great! Some of your examples include `library("forcis")` before reading example data and some don't...perhaps commit one way or the other.  I think it's usually pretty standard to exclude these for the package itself, but should be done you're attaching a dependency (like for ggplot2 in [this example](https://github.com/FRBCesab/forcis/blob/9926fdc8557e7f41a79664d8e0efcc18aadb9b4c/R/geom_basemap.R#L16))

-  quite a lot of `\dontrun{}`'s in the examples .. which could potentially be an issue if planning to submit to CRAN. Would any of the suggestions in [this chapter](https://r-pkgs.org/man.html#sec-man-examples-dependencies-conditional-execution) be appropriate to reduce these? the http and API complexities in your package examples/tests are far beyond the scope of anything I've ever written, so please disregard this feedback if it is irrelevant.

- I see you have added a `skip()` for the test that was causing failure in the earlier review, I did not get any failures on my end because of this! Though I get the same error when I run the test without it.

- All other tests pass; however, when running `devtools::test()` I get the message `"Relative counts from 24 samples could not be converted because of missing data on total assemblage"` quite a few times. Is it worth suppressing these to reduce clutter?
