+++ R/compute_abundances.R
@@ -2,93 +2,99 @@
 #' @export
 
 compute_abundances <- function(data, aggregate = TRUE) {
-  
   ## Check data ----
-  
+
   check_if_df(data)
-  
+
   if (get_data_type(data) %in% c("CPR North", "Sediment trap")) {
-    stop(paste0("This function is not designed to work with 'CPR North' or ", 
-                "'Sediment trap' data"), call. = FALSE) 
+    stop(
+      paste0(
+        "This function is not designed to work with 'CPR North' or ",
+        "'Sediment trap' data"
+      ),
+      call. = FALSE
+    )
   }
-  
+
   check_unique_taxonomy(data)
-  
-  
+
   taxa_cols <- get_species_names(data)
-  
 
   ## Raw data ----
 
   raw_data <- data[data$"subsample_count_type" == "Raw", ]
 
   raw_data <- tidyr::pivot_longer(
-    data      = raw_data, 
-    cols      = tidyr::all_of(taxa_cols), 
-    names_to  = "taxa", 
+    data = raw_data,
+    cols = tidyr::all_of(taxa_cols),
+    names_to = "taxa",
     values_to = "counts"
   )
 
   raw_data <- raw_data[!is.na(raw_data$"counts"), ]
 
   cols_to_remove <- c(
-    "subsample_count_type", "subsample_all_shells_present_were_counted", 
-    "total_of_forams_counted_ind", "sampling_device_type"
+    "subsample_count_type",
+    "subsample_all_shells_present_were_counted",
+    "total_of_forams_counted_ind",
+    "sampling_device_type"
   )
 
-  raw_data <- raw_data[ , !(colnames(raw_data) %in% cols_to_remove)]
+  raw_data <- raw_data[, !(colnames(raw_data) %in% cols_to_remove)]
 
-  colnames(raw_data)[grep("^counts$", colnames(raw_data))] <- 
+  colnames(raw_data)[grep("^counts$", colnames(raw_data))] <-
     "counts_raw_ab"
 
-  
   ## Absolute data ----
 
   abs_data <- data[data$"sample_volume_filtered" > 0, ]
   abs_data <- abs_data[abs_data$"subsample_count_type" == "Absolute", ]
 
   abs_data <- tidyr::pivot_longer(
-    data      = abs_data, 
-    cols      = tidyr::all_of(taxa_cols), 
-    names_to  = 'taxa', 
+    data = abs_data,
+    cols = tidyr::all_of(taxa_cols),
+    names_to = 'taxa',
     values_to = 'counts'
   )
 
   abs_data <- abs_data[!is.na(abs_data$"counts"), ]
 
-
   abs_data[["new_counts"]] <- floor(
     abs_data$"counts" * abs_data$"sample_volume_filtered"
   )
-  
-  cols_to_remove <- c("counts", "subsample_count_type", "sampling_device_type",
-    "subsample_all_shells_present_were_counted", "total_of_forams_counted_ind"
+
+  cols_to_remove <- c(
+    "counts",
+    "subsample_count_type",
+    "sampling_device_type",
+    "subsample_all_shells_present_were_counted",
+    "total_of_forams_counted_ind"
   )
 
-  abs_data <- abs_data[ , !(colnames(abs_data) %in% cols_to_remove)]
+  abs_data <- abs_data[, !(colnames(abs_data) %in% cols_to_remove)]
 
-  colnames(abs_data)[grep("^new_counts$", colnames(abs_data))] <- 
+  colnames(abs_data)[grep("^new_counts$", colnames(abs_data))] <-
     "counts_raw_ab"
-  
+
   abs_data <- abs_data[!duplicated(abs_data), ]
 
-
   ## Relative data ----
-  
+
   rel_data <- data[data$"sample_volume_filtered" > 0, ]
[…]
