+++ R/filter_by_bbox.R
@@ -1,96 +1,94 @@
-#' Filter FORCIS data by a spatial bounding box 
+#' Filter FORCIS data by a spatial bounding box
 #'
 #' @description
 #' Filters FORCIS data by a spatial bounding box.
-#' 
+#'
 #' @param data a `data.frame`. One obtained by `read_*_data()` functions.
-#' 
-#' @param bbox an object of class `bbox` (package `sf`) or a vector of four 
-#'   `numeric` values defining a square bounding box. Values must follow this 
-#'   order: minimum longitude (`xmin`), minimum latitude (`ymin`), maximum 
-#'   longitude (`xmax`), and maximum latitude (`ymax`). 
+#'
+#' @param bbox an object of class `bbox` (package `sf`) or a vector of four
+#'   `numeric` values defining a square bounding box. Values must follow this
+#'   order: minimum longitude (`xmin`), minimum latitude (`ymin`), maximum
+#'   longitude (`xmax`), and maximum latitude (`ymax`).
 #'   **Important:** if a vector of numeric values is provided, coordinates must
 #'   be defined in the system WGS 84 (`epsg=4326`).
 #'
-#' @return A `data.frame` containing a subset of `data` for the desired 
+#' @return A `data.frame` containing a subset of `data` for the desired
 #'   bounding box.
-#' 
+#'
 #' @export
 #'
 #' @examples
 #' # Attach the package ----
 #' library("forcis")
-#' 
+#'
 #' # Import example dataset ----
-#' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
+#' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 #'                          package = "forcis")
-#' 
+#'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
-#' 
+#'
 #' # Add 'data_type' column ----
 #' net_data$"data_type" <- "Net"
-#' 
+#'
 #' # Dimensions of the data.frame ----
 #' dim(net_data)
-#' 
+#'
 #' # Filter by oceans ----
 #' net_data_sub <- filter_by_bbox(net_data, bbox = c(45, -61, 82, -24))
-#' 
+#'
 #' # Dimensions of the data.frame ----
 #' dim(net_data_sub)
 
 filter_by_bbox <- function(data, bbox) {
-  
   if (missing(bbox)) {
     stop("Argument 'bbox' is required", call. = FALSE)
   }
-  
+
   data <- data[!is.na(data$"site_lon_start_decimal"), ]
   data <- data[!is.na(data$"site_lat_start_decimal"), ]
-  
+
   data_sf <- data_to_sf(data)
-  
-  
+
   ## Convert clip region into bbox object ----
-  
+
   if (inherits(bbox, "numeric")) {
-    
     if (length(bbox) != 4) {
-      stop("The object 'bbox' must a numeric of length 4 or a bbox object", 
-           call. = FALSE)
+      stop(
+        "The object 'bbox' must a numeric of length 4 or a bbox object",
+        call. = FALSE
+      )
     }
-    
-    bbox <- sf::st_bbox(c(xmin = bbox[1], ymin = bbox[2], 
-                          xmax = bbox[3], ymax = bbox[4]),
-                        crs = sf::st_crs(4326))  
+
+    bbox <- sf::st_bbox(
+      c(xmin = bbox[1], ymin = bbox[2], xmax = bbox[3], ymax = bbox[4]),
+      crs = sf::st_crs(4326)
+    )
   }
-  
-  
+
   ## Check bbox object ----
-  
+
[…]
