### 2025-03-28 7af8506
style: update function documentation
