+++ man/plot_record_by_season.Rd
@@ -20,7 +20,7 @@ This function produces a barplot of FORCIS sample records by season.
 library("forcis")
 
 # Import example dataset ----
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
+file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
 
 net_data <- read.table(file_name, dec = ".", sep = ";")
