+++ man/filter_by_ocean.Rd
@@ -23,7 +23,7 @@ Filters FORCIS data by one or several oceans.
 library("forcis")
 
 # Import example dataset ----
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
+file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
 
 net_data <- read.table(file_name, dec = ".", sep = ";")
