+++ man/select_forcis_columns.Rd
@@ -26,7 +26,7 @@ handle it and to speed up some computations.
 library("forcis")
 
 # Import example dataset ----
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
+file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
 
 net_data <- read.table(file_name, dec = ".", sep = ";")
