+++ tests/testthat/setup.R
@@ -1,20 +1,19 @@
 #' Setup Tests Infrastructure
-#' 
+#'
 
 ## Temporary Directory ----
 
 create_tempdir <- function(path = file.path(tempdir(), "sandbox")) {
-  
   old_wd <- getwd()
-  
+
   withr::defer(fs::dir_delete(path), envir = parent.frame())
-  
+
   dir.create(path)
-  
+
   setwd(path)
-  
+
   withr::defer(setwd(old_wd), envir = parent.frame())
-  
+
   invisible(path)
 }
 
