### 2025-03-28 3e62591
doc: suggest to use air in contributing
