### 2025-03-28 #660 @ahasverus
Hi @heyairf,
Thanks for your review and comments. I'm currently working on the comments raised by @sckott. After that, I will proceed with your review and improve the documentation.

Note that I have temporarily disabled unit tests for `download_forcis_db()` because I don't understand yet what the issue is (it has to do with the package `httptest2`). But I will figure it out.

Nicolas
