### 2025-03-28 f3faa3d
rOpenSci review: Format R code (#80)
* style: setup air formatter

* style: format r code

* style: update function documentation

* style: format r code in tests

* doc: suggest to use air in contributing
