+++ R/compute_concentrations.R
@@ -5,152 +5,158 @@
 #' relative abundance, and number concentration, using counts metadata.
 #'
 #' @param data a `data.frame`. One obtained by `read_*_data()` functions.
-#' 
+#'
 #' @param aggregate a `logical` of length 1. If `FALSE` counts will be derived
-#'   for each subsample. If `TRUE` (default) subsample counts will be 
+#'   for each subsample. If `TRUE` (default) subsample counts will be
 #'   aggregated by `sample_id`.
 #'
-#' @return A `data.frame` in long format with two additional columns: `taxa`, 
+#' @return A `data.frame` in long format with two additional columns: `taxa`,
 #' the taxon name and `counts_*`, the number concentration (`counts_n_conc`) or
-#' the relative abundance (`counts_rel_ab`) or the raw abundance 
+#' the relative abundance (`counts_rel_ab`) or the raw abundance
 #' (`counts_raw_ab`).
-#' 
+#'
 #' @details
-#' 
-#' - `compute_concentrations()` converts all counts to number concentrations 
+#'
+#' - `compute_concentrations()` converts all counts to number concentrations
 #' (n specimens/m³).
-#' - `compute_frequencies()` converts all counts to relative abundances 
+#' - `compute_frequencies()` converts all counts to relative abundances
 #' (% specimens per sampling unit).
-#' - `compute_abundances()` converts all counts to raw abundances 
+#' - `compute_abundances()` converts all counts to raw abundances
 #' (n specimens/sampling unit).
 #'
 #' @name computations
-#' 
+#'
 #' @examples
 #' # Import example dataset ----
-#' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
+#' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 #'                          package = "forcis")
-#' 
+#'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
-#' 
+#'
 #' # Select a taxonomy ----
 #' net_data <- select_taxonomy(net_data, taxonomy = "VT")
-#' 
+#'
 #' # Dimensions of the data.frame ----
 #' dim(net_data)
-#' 
+#'
 #' # Compute concentration ----
 #' net_data_conc <- compute_concentrations(net_data)
-#' 
+#'
 #' # Dimensions of the data.frame ----
 #' dim(net_data_conc)
 NULL
-
 
 
 #' @rdname computations
 #' @export
 
 compute_concentrations <- function(data, aggregate = TRUE) {
-  
   ## Check data ----
-  
+
   check_if_df(data)
-  
+
   if (get_data_type(data) == "Sediment trap") {
-    stop(paste0("This function is not designed to work with ", 
-                "'Sediment trap' data"), call. = FALSE) 
+    stop(
+      paste0(
+        "This function is not designed to work with ",
+        "'Sediment trap' data"
+      ),
+      call. = FALSE
+    )
   }
-  
+
   if (get_data_type(data) == "CPR North") {
-      
-    data[["min_conc_binned"]] <- 
+    data[["min_conc_binned"]] <-
       data$"count_bin_min" / data$"sample_volume_filtered"
 
-    data[["max_conc_binned"]] <- 
+    data[["max_conc_binned"]] <-
       data$"count_bin_max" / data$"sample_volume_filtered"
 
     cols_to_remove <- c("count_bin_min", "count_bin_max")
-  
[…]
