+++ tests/testthat/test-ggmap_data.R
@@ -7,11 +7,12 @@ colnames(df) <- get_required_columns()
 ## ggmap_data() ----
 
 test_that("Test ggmap_data() for success", {
-  
-  expect_silent({ gg <- ggmap_data(df) })
-  
+  expect_silent({
+    gg <- ggmap_data(df)
+  })
+
   expect_true("gg" %in% class(gg))
   expect_true("ggplot" %in% class(gg))
-  
+
   vdiffr::expect_doppelganger("Map data", gg)
 })

[…]
