### 2025-03-28 5eeaa0f
style: add prefix to global option in r code
