+++ R/read_plankton_nets_data.R
@@ -50,7 +50,7 @@ NULL
 read_plankton_nets_data <- function(
   path = ".",
   version = options()$"forcis_version",
-  check_for_update = options()$"check_for_update"
+  check_for_update = options()$"forcis_check_for_update"
 ) {
   ## Check args ----
 
