### 2025-03-28 50f7286
doc: update function documentations
