+++ man/download_forcis_db.Rd
@@ -7,7 +7,7 @@
 download_forcis_db(
   path = ".",
   version = options()$forcis_version,
-  check_for_update = options()$check_for_update,
+  check_for_update = options()$forcis_check_for_update,
   overwrite = FALSE,
   timeout = 60
 )
@@ -26,11 +26,11 @@ for the current R session. Note that it is recommended to use the latest
 version of the database.}
 
 \item{check_for_update}{a \code{logical}. If \code{TRUE} (default) the function will
-check if a newer version of the FORCIS database is available on Zenodo and
-will print an informative message. Note that this argument can be handle
-with the global option \code{check_for_update}. For example, if user calls
-\code{options(check_for_update = FALSE)}, the message to download the latest
-version will be disabled for the current R session.}
+check if a newer version of the FORCIS database is available on Zenodo
+and will print an informative message. Note that this argument can be
+handle with the global option \code{forcis_check_for_update}. For example, if
+user calls \code{options(forcis_check_for_update = FALSE)}, the message to
+download the latest version will be disabled for the current R session.}
 
 \item{overwrite}{a \code{logical}. If \code{TRUE} it will override the downloaded
 files of the FORCIS database. Default is \code{FALSE}.}
