### 2025-03-28 efd3b2a
rOpenSci review: Rename global options (#81)
* style: add prefix to global option in r code

* doc: update function documentations

* doc: add prefix to global option in vignette
