### 2025-03-28 0f85fb7
feat: functions always return tibble
