+++ R/read_sediment_trap_data.R
@@ -79,5 +79,5 @@ read_sediment_trap_data <- function(
     data[, taxa_columns[i]] <- as.numeric(data[, taxa_columns[i]])
   }
 
-  data
+  tibble::as_tibble(data)
 }
