### 2025-03-28 ab5199f
doc: update function input/output
