### 2025-03-28 7af13e7
test: adapt tests to handle tibble
