### 2025-03-28 cbaaae6
rOpenSci review: Homogeneize function output (#82)
* feat: functions always return tibble

* doc: update function input/output

* test: adapt tests to handle tibble
