+++ R/read_cpr_south_data.R
@@ -79,5 +79,5 @@ read_cpr_south_data <- function(
     data[, taxa_columns[i]] <- as.numeric(data[, taxa_columns[i]])
   }
 
-  data
+  tibble::as_tibble(data)
 }
