### 2025-04-01 #660 @mpadge
@beatrizmilz In case you didn't realise, you still need to log Scott's review, as explained in [*Guide for Editors* chapter](https://devguide.ropensci.org/softwarereview_editor.html#during-review), in the `submit review` bit 😄
