### 2025-04-03 01845aa
chore: clean up download mock one call to zenodo api is enough for other calls
