### 2025-04-03 6c0ac8b
refactor(wip): change zenodo endpoints The behavior of downloading files should change: 1. check metadata of latest version 2. use a path like cesab/forcis/data/{version}/{files} you may use the metadat to check if version dir exist and if the files exists or not and verify withchecksum. 3. download only the missing files. 4. user can specify a version: use the endpoint to check if it exist 1st then do the same as 2. & 3. 5. expose a function to return available versions to make it easier to select one of them 6. expose a function to clean up all forcis data under cesab/forcis folder
