### 2025-04-03 3b39419
chore: Fix some tests * The second call inside with_dir_mock may cause the test to fail if run a second time. Therefore, either delete the mock file or separate them. * TODO: The method for checking version(s) needs to be updated. You can retrieve the latest version via a specific Zenedo endpoint or by requesting a list of versions (ensure the page size is large enough to include all versions in a single response).
