### 2025-04-03 05cc9a6
refactor & fix test get_version_metadata TODO: extract helper fucntions
