+++ R/get_version_metadata.R
@@ -22,16 +22,40 @@
 #' get_version_metadata()
 
 get_version_metadata <- function(version = NULL) {
-  ## Check arguments ----
-
   check_version(version)
+  
+  # Retrieve metadata
+  res <- get_metadata(version)
+  
+  # Extract versions and determine position
+  if (version == "latest") {
+    versions <- extract_single_version(res)
+    pos <- 1  # Since it's a single item
+  } else {
+    versions <- extract_versions(res)
+    pos <- determine_version_position(version, versions)
+  }
+  
+  # Extract metadata and files
+  meta <- extract_metadata(res, pos, version)
+  files <- extract_files(res, pos, version)
+  
+  meta$"files" <- files
+  meta$"resource_type" <- meta$"resource_type"$"type"
+  
+  meta <- clean_metadata(meta)
+  meta
+}
 
-  ## Retrieve information ----
+extract_single_version <- function(res) {
+  data.frame(
+    "publication_date" = res$"metadata"$"publication_date",
+    "version" = res$"metadata"$"version",
+    "access_right" = res$"metadata"$"access_right"
+  )
+}
 
-  res <- get_metadata(version = version)
-
-  ## Get all versions information ----
-
+extract_versions <- function(res) {
   versions <- lapply(res$"hits"$"hits", function(x) {
     data.frame(
       "publication_date" = x$"metadata"$"publication_date",
@@ -39,46 +63,57 @@ get_version_metadata <- function(version = NULL) {
       "access_right" = x$"metadata"$"access_right"
     )
   })
+  do.call(rbind.data.frame, versions)
+}
 
-  versions <- do.call(rbind.data.frame, versions)
-
-  ## Subset version ----
-
+determine_version_position <- function(version, versions) {
   if (is.null(version) || version == "latest") {
     pos <- which.max(as.Date(versions$"publication_date"))
   } else {
     pos <- which(versions$"version" == version)
-
     if (length(pos) == 0) {
-      stop(
-        "The required version is not available. Please run ",
-        "'get_available_versions()' to list available versions."
-      )
+      stop("The required version is not available. Please run 'get_available_versions()' to list available versions.")
     }
   }
+  
+  if (length(pos) == 0) {
+    stop("No valid version found in the metadata.")
+  }
+  
+  pos
+}
 
-  meta <- res$"hits"$"hits"$"metadata"
-  files <- res$"hits"$"hits"$"files"
-
-  meta <- lapply(res$"hits"$"hits", function(x) x$"metadata")
-  meta <- meta[[pos]]
-
+extract_metadata <- function(res, pos, version) {
+  if (version == "latest") {
+    meta <- res$"metadata"
+  } else {
+    meta <- lapply(res$"hits"$"hits", function(x) x$"metadata")
+    meta <- meta[[pos]]
+  }
   meta$"keywords" <- unlist(meta$"keywords")
   meta$"license" <- unlist(meta$"license"$"id")
+  meta$"creators" <- extract_creators(meta$"creators")
+  meta
+}
 
-  meta$"creators" <- lapply(meta$"creators", function(x) {
+extract_creators <- function(creators) {
+  creators <- lapply(creators, function(x) {
     data.frame(
       "name" = x$name,
       "affiliation" = ifelse(is.null(x$affiliation), NA, x$affiliation),
       "orcid" = ifelse(is.null(x$orcid), NA, x$orcid)
     )
   })
+  do.call(rbind.data.frame, creators)
+}
 
-  meta$"creators" <- do.call(rbind.data.frame, meta$"creators")
-
[…]
