+++ R/get_version_metadata.R
@@ -5,7 +5,7 @@
 #' the Zenodo API (\url{https://developers.zenodo.org}).
 #'
 #' @param version a `character` of length 1. The label of the version. Use
-#'   [get_available_versions()] to list available versions. If `NULL` (default)
+#'   [get_available_versions()] to list available versions. If `latest` (default)
 #'   the latest version is used.
 #'
 #' @return A `list` with all information about the version, including: `title`,
@@ -21,28 +21,28 @@
 #' # Get information for the latest version of the FORCIS database ----
 #' get_version_metadata()
 
-get_version_metadata <- function(version = NULL) {
+get_version_metadata <- function(version = "latest") {
   check_version(version)
-  
+
   # Retrieve metadata
   res <- get_metadata(version)
-  
+
   # Extract versions and determine position
   if (version == "latest") {
     versions <- extract_single_version(res)
-    pos <- 1  # Since it's a single item
+    pos <- 1 # Since it's a single item
   } else {
     versions <- extract_versions(res)
     pos <- determine_version_position(version, versions)
   }
-  
+
   # Extract metadata and files
   meta <- extract_metadata(res, pos, version)
   files <- extract_files(res, pos, version)
-  
+
   meta$"files" <- files
   meta$"resource_type" <- meta$"resource_type"$"type"
-  
+
   meta <- clean_metadata(meta)
   meta
 }
@@ -72,14 +72,16 @@ determine_version_position <- function(version, versions) {
   } else {
     pos <- which(versions$"version" == version)
     if (length(pos) == 0) {
-      stop("The required version is not available. Please run 'get_available_versions()' to list available versions.")
+      stop(
+        "The required version is not available. Please run 'get_available_versions()' to list available versions."
+      )
     }
   }
-  
+
   if (length(pos) == 0) {
     stop("No valid version found in the metadata.")
   }
-  
+
   pos
 }
 
