### 2025-04-04 6ffaebf
Refactor internals
