+++ R/internal_taxonomy.R
@@ -1,0 +1,383 @@
+#' Species names and taxonomy name
+#'
+#' @noRd
+
+species_list <- function() {
+  data.frame(
+    "taxon" = c(
+      "b_digitata_LT",
+      "b_pumilio_LT",
+      "b_variabilis_LT",
+      "c_nitida_LT",
+      "d_anfracta_LT",
+      "g_adamsi_LT",
+      "g_bulloides_LT",
+      "g_calida_LT",
+      "g_cavernula_LT",
+      "g_conglobatus_LT",
+      "g_conglomerata_LT",
+      "g_crassaformis_LT",
+      "g_falconensis_LT",
+      "g_glutinata_LT",
+      "g_hexagona_LT",
+      "g_hirsuta_LT",
+      "g_inflata_LT",
+      "g_cultrata_LT",
+      "g_minuta_LT",
+      "g_ruber_any_LT",
+      "g_rubescens_LT",
+      "g_scitula_LT",
+      "g_siphonifera_LT",
+      "g_tenellus_LT",
+      "g_theyeri_LT",
+      "g_truncatulinoides_LT",
+      "g_tumida_LT",
+      "g_ungulata_LT",
+      "g_uvula_LT",
+      "n_vivans_LT",
+      "h_digitata_LT",
+      "h_pelagica_LT",
+      "n_dutertrei_LT",
+      "n_incompta_LT",
+      "n_pachyderma_any_LT",
+      "o_riedeli_LT",
+      "o_universa_LT",
+      "p_obliquiloculata_LT",
+      "s_dehiscens_LT",
+      "t_clarkei_LT",
+      "t_fleisheri_LT",
+      "t_humilis_LT",
+      "t_iota_LT",
+      "t_parkerae_LT",
+      "t_quinqueloba_LT",
+      "t_sacculifer_LT",
+      "UniD_LT",
+
+      "b_digitata_VT",
+      "b_pumilio_VT",
+      "b_variabilis_VT",
+      "c_nitida_VT",
+      "d_anfracta_VT",
+      "g_adamsi_VT",
+      "g_bulloides_VT",
+      "g_calida_VT",
+      "g_cavernula_VT",
+      "g_conglobatus_VT",
+      "g_conglomerata_VT",
+      "g_crassaformis_VT",
+      "g_elongatus_VT",
+      "g_falconensis_VT",
+      "g_glutinata_VT",
+      "g_hexagona_VT",
+      "g_hirsuta_VT",
+      "g_inflata_VT",
+      "g_cultrata_VT",
+      "g_minuta_VT",
+      "g_ruber_albus_VT",
+      "g_ruber_albus_or_elongatus_VT",
+      "g_ruber_any_VT",
+      "g_ruber_ruber_VT",
+      "g_rubescens_VT",
+      "g_scitula_VT",
+      "g_siphonifera_VT",
+      "g_tenellus_VT",
+      "g_theyeri_VT",
+      "g_truncatulinoides_VT",
+      "g_truncatulinoides_left_VT",
+      "g_truncatulinoides_right_VT",
+      "g_tumida_VT",
+      "g_ungulata_VT",
+      "g_uvula_VT",
+      "n_vivans_VT",
+      "h_digitata_VT",
+      "h_pelagica_VT",
+      "n_dutertrei_VT",
+      "n_incompta_VT",
+      "n_pachyderma_VT",
+      "n_pachyderma_incompta_VT",
+      "o_riedeli_VT",
+      "o_universa_VT",
+      "p_obliquiloculata_VT",
+      "s_dehiscens_VT",
+      "t_clarkei_VT",
+      "t_fleisheri_VT",
+      "t_humilis_VT",
+      "t_iota_VT",
[…]

[…]
