### 2025-04-04 0f76309
chores: code clean up and format
