### 2025-04-04 06f0e39
style: use lintr & styler to enforce tidyverse style
