+++ R/compute_concentrations.R
@@ -30,7 +30,8 @@
 #' @examples
 #' # Import example dataset ----
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
-#'                          package = "forcis")
+#'   package = "forcis"
+#' )
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #'
@@ -91,8 +92,8 @@ compute_concentrations <- function(data, aggregate = TRUE) {
   abs_data <- tidyr::pivot_longer(
     data = abs_data,
     cols = tidyr::all_of(taxa_cols),
-    names_to = 'taxa',
-    values_to = 'counts'
+    names_to = "taxa",
+    values_to = "counts"
   )
 
   abs_data <- abs_data[!is.na(abs_data$"counts"), ]
@@ -149,8 +150,8 @@ compute_concentrations <- function(data, aggregate = TRUE) {
   rel_data <- tidyr::pivot_longer(
     data = rel_data,
     cols = tidyr::all_of(taxa_cols),
-    names_to = 'taxa',
-    values_to = 'counts'
+    names_to = "taxa",
+    values_to = "counts"
   )
 
   rel_data <- rel_data[!is.na(rel_data$"counts"), ]
