+++ R/get_available_versions.R
@@ -17,7 +17,6 @@
 #'
 #' # Versions of the FORCIS database ----
 #' get_available_versions()
-
 get_available_versions <- function() {
   ## Retrieve information ----
 
