+++ R/plot_record_by_month.R
@@ -15,13 +15,13 @@
 #'
 #' # Import example dataset ----
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
-#'                          package = "forcis")
+#'   package = "forcis"
+#' )
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #'
 #' # Plot data by year (example dataset) ----
 #' plot_record_by_month(net_data)
-
 plot_record_by_month <- function(data) {
   ## Check data object ----
 
