+++ R/select_forcis_columns.R
@@ -22,7 +22,8 @@
 #'
 #' # Import example dataset ----
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
-#'                          package = "forcis")
+#'   package = "forcis"
+#' )
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #'
@@ -43,7 +44,6 @@
 #'
 #' # Dimensions of the data.frame ----
 #' dim(net_data)
-
 select_forcis_columns <- function(data, cols = NULL) {
   # Check args ----
 
