+++ R/filter_by_month.R
@@ -18,7 +18,8 @@
 #'
 #' # Import example dataset ----
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
-#'                          package = "forcis")
+#'   package = "forcis"
+#' )
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #'
@@ -33,7 +34,6 @@
 #'
 #' # Dimensions of the data.frame ----
 #' dim(net_data_sub)
-
 filter_by_month <- function(data, months) {
   ## Check data object ----
 
