+++ man/get_version_metadata.Rd
@@ -8,8 +8,8 @@ get_version_metadata(version = "latest")
 }
 \arguments{
 \item{version}{a \code{character} of length 1. The label of the version. Use
-\code{\link[=get_available_versions]{get_available_versions()}} to list available versions. If \code{latest} (default)
-the latest version is used.}
+\code{\link[=get_available_versions]{get_available_versions()}} to list available versions.
+If \code{latest} (default) the latest version is used.}
 }
 \value{
 A \code{list} with all information about the version, including: \code{title},
