### 2025-04-07 b295b29
refactor: read functions squash
