+++ tests/testthat/test-read_plankton_nets_data.R
@@ -23,7 +23,7 @@ test_that("Test read_plankton_nets_data() for error", {
   expect_error(
     read_plankton_nets_data(version = "07"),
     paste0(
-      "The Plankton net dataset does not exist. Please use ",
+      "The Net dataset does not exist. Please use ",
       "the function 'download_forcis_db()'."
     ),
     fixed = TRUE
