### 2025-04-08 3478141
feat: add md5 hecksum validation
