### 2025-04-09 fc4de7e
feat: add cache helper functions
