+++ DESCRIPTION
@@ -66,5 +66,6 @@ Suggests:
     vdiffr,
     withr,
     lintr,
-    styler
+    styler,
+    rappdirs
 Config/testthat/edition: 3
