### 2025-04-09 dce06ed
feat: refactor get_data_dir
