### 2025-04-09 410a1cb
feat: add sanitize_version function to clean version strings for directory names
