### 2025-04-09 23e5c74
fix: update parameter name from 'replacement' to 'separator' in clean_cache documentation
