### 2025-04-09 2520bab
chore: add rappdirs to Imports instead of Suggests
