+++ DESCRIPTION
@@ -56,7 +56,8 @@ Imports:
     tibble,
     tidyr,
     utils,
-    vroom
+    vroom,
+    rappdirs
 Suggests: 
     fs,
     httptest2,
@@ -66,6 +67,5 @@ Suggests:
     vdiffr,
     withr,
     lintr,
-    styler,
-    rappdirs
+    styler
 Config/testthat/edition: 3
