+++ R/read_plankton_nets_data.R
@@ -50,7 +50,7 @@ read_plankton_nets_data <- function(
     path = ".",
     version = options()$"forcis_version",
     check_for_update = options()$"forcis_check_for_update") {
-  read_dataset_by_name(
+  load_forcis(
     name = "net",
     version = version,
     check_for_update = check_for_update
