+++ R/read_cpr_north_data.R
@@ -5,7 +5,7 @@ read_cpr_north_data <- function(
     path = ".",
     version = options()$"forcis_version",
     check_for_update = options()$"forcis_check_for_update") {
-  read_dataset_by_name(
+  load_forcis(
     name = "cpr_north",
     version = version,
     check_for_update = check_for_update
