+++ R/read_pump_data.R
@@ -5,7 +5,7 @@ read_pump_data <- function(
     path = ".",
     version = options()$"forcis_version",
     check_for_update = options()$"forcis_check_for_update") {
-  read_dataset_by_name(
+  load_forcis(
     name = "pump",
     version = version,
     check_for_update = check_for_update
