+++ R/read_pump_data.R
@@ -1,6 +1,7 @@
 #' @rdname read_data
 #' @export
 
+# TODO: to be removed & replaced by load_forcis("pump")
 read_pump_data <- function(
     path = ".",
     version = options()$"forcis_version",
