+++ R/read_cpr_south_data.R
@@ -1,6 +1,7 @@
 #' @rdname read_data
 #' @export
 
+# TODO: to be removed & replaced by load_forcis("cpr_south")
 read_cpr_south_data <- function(
     path = ".",
     version = options()$"forcis_version",
