### 2025-04-09 24dc4a4
fix: replace sapply by vapply suggestion by goodpractice pkg
