### 2025-04-09 ba25fb5
chore: use backports instead of rappdirs cran restriction for rappdirs
