### 2025-04-09 0bea47b
fix: correct the use of R_user_dir
