+++ man/load_forcis.Rd
@@ -1,0 +1,32 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/load_forcis.R
+\name{load_forcis}
+\alias{load_forcis}
+\title{Load a FORCIS dataset by name}
+\usage{
+load_forcis(
+  name,
+  path = ".",
+  version = options()$forcis_version,
+  check_for_update = options()$forcis_check_for_update,
+  cached = TRUE
+)
+}
+\arguments{
+\item{name}{Key identifying the dataset ("net", "pump", etc.)}
+
+\item{path}{The directory path to read from}
+
+\item{version}{Database version to use (TODO: should be NULL by default)}
+
+\item{check_for_update}{Whether to check for database updates
+(TODO: to be removed)}
+
+\item{cached}{TODO: for later use (cache dataset on local storage)}
+}
+\value{
+A tibble containing the dataset
+}
+\description{
+Load a FORCIS dataset by name
+}
