### 2025-04-10 e3adca7
refactor: files check functions
