### 2025-04-10 dd3f730
feat: add more helper functions * get_option() * set_option() * log_message()
