### 2025-04-10 7101119
test(with_mock_dir): switch simplify to FALSE
