+++ tests/testthat/mockdata/v999/zenodo.org/api/records-ecb41d.R
@@ -1,0 +1,16 @@
+structure(list(method = "GET", url = "https://zenodo.org/api/records?q=conceptrecid%3A7390791%20AND%20version%3A999&all_versions=true&size=999", 
+    status_code = 200L, headers = structure(list(server = "nginx", 
+        date = "Thu, 10 Apr 2025 10:01:39 GMT", `content-type` = "application/json", 
+        `transfer-encoding` = "chunked", vary = "Accept-Encoding", 
+        `x-ratelimit-limit` = "133", `x-ratelimit-remaining` = "56", 
+        `x-ratelimit-reset` = "1744279333", `retry-after` = "33", 
+        `permissions-policy` = "interest-cohort=()", `x-frame-options` = "sameorigin", 
+        `x-xss-protection` = "1; mode=block", `x-content-type-options` = "nosniff", 
+        `content-security-policy` = "default-src 'self' fonts.googleapis.com *.gstatic.com data: 'unsafe-inline' 'unsafe-eval' blob: zenodo-broker.web.cern.ch zenodo-broker-qa.web.cern.ch maxcdn.bootstrapcdn.com cdnjs.cloudflare.com ajax.googleapis.com webanalytics.web.cern.ch", 
+        `strict-transport-security` = "max-age=31556926; includeSubDomains", 
+        `referrer-policy` = "strict-origin-when-cross-origin", 
+        `access-control-allow-origin` = "*", `access-control-expose-headers` = "Content-Type, ETag, Link, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset", 
+        `strict-transport-security` = "max-age=15768000", `x-request-id` = "63f76824e99f46d7901bbe2bd58b7f77", 
+        `content-encoding` = "gzip", `set-cookie` = "REDACTED"), redact = character(0), class = "httr2_headers"), 
+    body = charToRaw("{\"hits\": {\"hits\": [], \"total\": 0}, \"aggregations\": {\"access_status\": {\"buckets\": [], \"label\": \"Access status\"}, \"resource_type\": {\"buckets\": [], \"label\": \"Resource types\"}, \"subject\": {\"buckets\": [], \"label\": \"Subjects\"}, \"file_type\": {\"buckets\": [], \"label\": \"File type\"}}, \"links\": {\"self\": \"https://zenodo.org/api/records?allversions=True&page=1&q=conceptrecid%3A7390791%20AND%20version%3A999&size=999&sort=bestmatch\"}}"), 
+    cache = new.env(parent = emptyenv())), class = "httr2_response")
