+++ tests/testthat/mockdata/all/zenodo.org/api/records/12724286/versions-4eba17.R
@@ -1,0 +1,16 @@
+structure(list(method = "GET", url = "https://zenodo.org/api/records/12724286/versions?size=999&sort=version", 
+    status_code = 200L, headers = structure(list(server = "nginx", 
+        date = "Thu, 10 Apr 2025 10:01:38 GMT", `content-type` = "application/json", 
+        `transfer-encoding` = "chunked", vary = "Accept-Encoding", 
+        `access-control-allow-origin` = "*", `access-control-expose-headers` = "Content-Type, ETag, Link, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset", 
+        `x-ratelimit-limit` = "133", `x-ratelimit-remaining` = "57", 
+        `x-ratelimit-reset` = "1744279333", `retry-after` = "34", 
+        `permissions-policy` = "interest-cohort=()", `x-frame-options` = "sameorigin", 
+        `x-xss-protection` = "1; mode=block", `x-content-type-options` = "nosniff", 
+        `content-security-policy` = "default-src 'self' fonts.googleapis.com *.gstatic.com data: 'unsafe-inline' 'unsafe-eval' blob: zenodo-broker.web.cern.ch zenodo-broker-qa.web.cern.ch maxcdn.bootstrapcdn.com cdnjs.cloudflare.com ajax.googleapis.com webanalytics.web.cern.ch", 
+        `strict-transport-security` = "max-age=31556926; includeSubDomains", 
+        `referrer-policy` = "strict-origin-when-cross-origin", 
+        `strict-transport-security` = "max-age=15768000", `x-request-id` = "cf239351f7294ead62f5542683db26cc", 
+        `content-encoding` = "gzip", `set-cookie` = "REDACTED"), redact = character(0), class = "httr2_headers"), 
[…]
