+++ tests/testthat/mockdata/latest/zenodo.org/api/records/12724286/versions/latest.R
@@ -1,0 +1,17 @@
+structure(list(method = "GET", url = "https://zenodo.org/api/records/12724286", 
+    status_code = 200L, headers = structure(list(server = "nginx", 
+        date = "Thu, 10 Apr 2025 10:01:39 GMT", `content-type` = "application/json", 
+        `transfer-encoding` = "chunked", vary = "Accept-Encoding", 
+        etag = "W/\"4\"", link = "<https://zenodo.org/api/records/12724286> ; rel=\"linkset\" ; type=\"application/linkset+json\"", 
+        `x-ratelimit-limit` = "133", `x-ratelimit-remaining` = "54", 
+        `x-ratelimit-reset` = "1744279333", `retry-after` = "33", 
+        `permissions-policy` = "interest-cohort=()", `x-frame-options` = "sameorigin", 
+        `x-xss-protection` = "1; mode=block", `x-content-type-options` = "nosniff", 
+        `content-security-policy` = "default-src 'self' fonts.googleapis.com *.gstatic.com data: 'unsafe-inline' 'unsafe-eval' blob: zenodo-broker.web.cern.ch zenodo-broker-qa.web.cern.ch maxcdn.bootstrapcdn.com cdnjs.cloudflare.com ajax.googleapis.com webanalytics.web.cern.ch", 
+        `strict-transport-security` = "max-age=31556926; includeSubDomains", 
+        `referrer-policy` = "strict-origin-when-cross-origin", 
+        `access-control-allow-origin` = "*", `access-control-expose-headers` = "Content-Type, ETag, Link, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset", 
+        `strict-transport-security` = "max-age=15768000", `x-request-id` = "527631ff7321979a9e5ee82536b9eb2e", 
+        `content-encoding` = "gzip"), redact = character(0), class = "httr2_headers"), 
[…]
