+++ man/load_forcis.Rd
@@ -30,3 +30,4 @@ A tibble containing the dataset
 \description{
 Load a FORCIS dataset by name
 }
+\keyword{internal}
