+++ NAMESPACE
@@ -20,7 +20,6 @@ export(get_required_columns)
 export(get_species_names)
 export(get_version_metadata)
 export(ggmap_data)
-export(load_forcis)
 export(plot_record_by_depth)
 export(plot_record_by_month)
 export(plot_record_by_season)
