+++ CONTRIBUTING.md
@@ -2,13 +2,11 @@
 
 First off, thanks for taking the time to contribute to `forcis`!
 
-All types of contributions are encouraged and valued. See the 
-[Table of contents](#table-of-contents) for different ways to help and details 
-about how this project handles them. Please make sure to read the relevant 
-section before making your contribution. It will make it a lot easier for us 
+All types of contributions are encouraged and valued. See the
+[Table of contents](#table-of-contents) for different ways to help and details
+about how this project handles them. Please make sure to read the relevant
+section before making your contribution. It will make it a lot easier for us
 maintainers and smooth out the experience for all involved.
-
-
 
 ## Table of contents
 
@@ -20,43 +18,57 @@ maintainers and smooth out the experience for all involved.
 - [Requesting features](#requesting-features)
 - [Contributing code](#contributing-code)
 
-
-
 ## Code of conduct
 
 This project is released with a
 [Contributor Code of Conduct](https://github.com/frbcesab/forcis/blob/master/CODE_OF_CONDUCT.md).
-By participating, you are expected to uphold this code. Please report 
+By participating, you are expected to uphold this code. Please report
 unacceptable behavior to <nicolas.casajus@fondationbiodiversite.fr>.
-
-
 
 ## Style guide
 
-We use the [Tidyverse style guide](https://style.tidyverse.org/) for writing R 
-code. You can install [styler](https://github.com/r-lib/styler/) to automatically 
-format your R code. Functions are documented with the 
-[roxygen2](https://roxygen2.r-lib.org/articles/roxygen2.html) syntax. 
-`forcis` uses the `lower_snake_case`.
+We use the [Tidyverse style guide](https://style.tidyverse.org/) for writing R code. To ensure code consistency:
 
+- Use [styler](https://github.com/r-lib/styler/) to automatically format your R code
+- Use [lintr](https://github.com/r-lib/lintr) to identify style and possible syntax issues
+- Use [goodpractice](https://github.com/mangothecat/goodpractice) to ensure rOpenSci best practices
 
+Before submitting a PR, please run these checks:
+
+```r
+styler::style_pkg()
+lintr::lint_package()
+goodpractice::gp()
+```
+
+Functions are documented with the
+[roxygen2](https://roxygen2.r-lib.org/articles/roxygen2.html) syntax.
+
+`forcis` uses the `lower_snake_case` naming convention.
 
 ## Commit messages
 
-If you want to contribute by commiting changes, please try to use the 
-[Conventional commits](https://www.conventionalcommits.org/en/v1.0.0/) 
+If you want to contribute by commiting changes, please try to use the
+[Conventional commits](https://www.conventionalcommits.org/en/v1.0.0/)
 specification.
 
+### Setting up pre-commit hooks
 
+We recommend using [pre-commit](https://pre-commit.com/) to automatically check your code before committing. This ensures style consistency and prevents committing code with style issues.
+
+1. Install pre-commit: `pip install pre-commit`
+2. Install the hooks: `pre-commit install`
+
+The repository includes a `.pre-commit-config.yaml` file configured to run R-specific hooks including style checks.
 
 ## Asking questions
 
 Before you ask a question, it is best to search for existing
-[Issues](https://github.com/frbcesab/forcis/issues) that might help 
-you. In case you have found a suitable issue and still need clarification, you 
+[Issues](https://github.com/frbcesab/forcis/issues) that might help
+you. In case you have found a suitable issue and still need clarification, you
 can write your question in this issue.
 
-If you then still feel the need to ask a question and need clarification, we 
+If you then still feel the need to ask a question and need clarification, we
 recommend the following:
 
[…]
