### 2025-04-10 48c1192
chore: add pre-commit configuration (empty)
