### 2025-04-11 e2d90c1
chore(pre-commit): update configuration
