+++ .pre-commit-config.yaml
@@ -1,16 +1,20 @@
 # All available hooks: https://pre-commit.com/hooks.html
 # R specific hooks: https://github.com/lorenzwalthert/precommit
 repos:
--   repo: https://github.com/lorenzwalthert/precommit
+  - repo: https://github.com/lorenzwalthert/precommit
     rev: v0.4.3.9008
     hooks:
-    -   id: style-files
-        args: [--style_pkg=styler, --style_fun=tidyverse_style]    
-    -   id: roxygenize
-    # codemeta must be above use-tidy-description when both are used
-    # -   id: codemeta-description-updated
-    -   id: use-tidy-description
-    -   id: spell-check
+      - id: style-files
+        args: [--style_pkg=styler, --style_fun=tidyverse_style]
+      - id: roxygenize
+        additional_dependencies:
+          - dplyr
+          - roxygen2
+          - rlang
+      # codemeta must be above use-tidy-description when both are used
+      # -   id: codemeta-description-updated
+      - id: use-tidy-description
+      - id: spell-check
         exclude: >
           (?x)^(
           .*\.[rR]|
@@ -40,31 +44,36 @@ repos:
           \.github/workflows/.*|
           data/.*|
           )$
-    -   id: lintr
-    -   id: readme-rmd-rendered
-    -   id: parsable-R
-    -   id: no-browser-statement
-    -   id: no-print-statement
-    -   id: no-debug-statement
-    -   id: deps-in-desc
-    -   id: pkgdown
--   repo: https://github.com/pre-commit/pre-commit-hooks
+      - id: lintr
+        additional_dependencies:
+          - r-lib/lintr
+          - any::cyclocomp
+      - id: readme-rmd-rendered
+      - id: parsable-R
+      - id: no-browser-statement
+      - id: no-print-statement
+      - id: no-debug-statement
+      - id: deps-in-desc
+      - id: pkgdown
+  - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v5.0.0
-    hooks: 
-    -   id: check-added-large-files
-        args: ['--maxkb=200']
-    -   id: file-contents-sorter
+    hooks:
+      - id: check-added-large-files
+        args: ["--maxkb=200"]
+      - id: check-yaml
+      - id: mixed-line-ending
+      - id: file-contents-sorter
         files: '^\.Rbuildignore$'
-    -   id: end-of-file-fixer
+      - id: end-of-file-fixer
         exclude: '\.Rd'
--   repo: https://github.com/pre-commit-ci/pre-commit-ci-config
+  - repo: https://github.com/pre-commit-ci/pre-commit-ci-config
     rev: v1.6.1
     hooks:
-    # Only required when https://pre-commit.ci is used for config validation
-    -   id: check-pre-commit-ci-config
--   repo: local
+      # Only required when https://pre-commit.ci is used for config validation
+      - id: check-pre-commit-ci-config
+  - repo: local
     hooks:
-    -   id: forbid-to-commit
+      - id: forbid-to-commit
         name: Don't commit common R artifacts
         entry: Cannot commit .Rhistory, .RData, .Rds or .rds.
         language: fail
@@ -72,5 +81,5 @@ repos:
         # `exclude: <regex>` to allow committing specific files
 
 ci:
-    autoupdate_schedule: monthly
-    skip: [pkgdown]
+  autoupdate_schedule: monthly
+  skip: [pkgdown]
