### 2025-04-11 7c67173
docs(CONTRIBUTING): clarify optional pre-commit hooks section
