### 2025-04-11 d2aa323
fix: correct options structure in set_option
