+++ .pre-commit-config.yaml
@@ -11,6 +11,11 @@ repos:
           - dplyr
           - roxygen2
           - rlang
+          - ggplot2
+          - httr2
+          - sf
+          - tidyr
+          - vroom
       # codemeta must be above use-tidy-description when both are used
       # -   id: codemeta-description-updated
       - id: use-tidy-description
