### 2025-04-11 a90771f
feat: enhance get_latest_version
