+++ .pre-commit-config.yaml
@@ -7,15 +7,6 @@ repos:
       - id: style-files
         args: [--style_pkg=styler, --style_fun=tidyverse_style]
       - id: roxygenize
-        additional_dependencies:
-          - dplyr
-          - roxygen2
-          - rlang
-          - ggplot2
-          - httr2
-          - sf
-          - tidyr
-          - vroom
       # codemeta must be above use-tidy-description when both are used
       # -   id: codemeta-description-updated
       - id: use-tidy-description
@@ -50,9 +41,6 @@ repos:
           data/.*|
           )$
       - id: lintr
-        additional_dependencies:
-          - r-lib/lintr
-          - any::cyclocomp
       - id: readme-rmd-rendered
       - id: parsable-R
       - id: no-browser-statement
