### 2025-04-11 2db5087
style(lintr): fix some styles
