### 2025-04-11 1a01571
chore(precommit): tweak lintr deps
