+++ .pre-commit-config.yaml
@@ -53,7 +53,6 @@ repos:
       - id: lintr
         args: [--load_package]
         additional_dependencies:
-          - r-lib/lintr
           - ggplot2
           - httr2
           - rlang
