### 2025-04-14 220ae45
feat: Add Zenodo metadata handling and file download
