### 2025-04-15 a37875d
feat: improve caching in load_forcis_test
