### 2025-04-15 de432e7
docs(get_meta_cache_path): fix a typo
