### 2025-04-15 e9bc290
fix: restricted version & cache bug
