+++ man/print_available_versions.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/print_available_versions.R
+\name{print_available_versions}
+\alias{print_available_versions}
+\title{Print Available Versions}
+\usage{
+print_available_versions()
+}
+\value{
+Invisibly returns NULL, but prints output to console.
+}
+\description{
+Prints a human-readable formatted list of all available versions.
+}
+\examples{
+print_available_versions()
+
+}
