+++ NAMESPACE
@@ -24,6 +24,7 @@ export(plot_record_by_depth)
 export(plot_record_by_month)
 export(plot_record_by_season)
 export(plot_record_by_year)
+export(print_available_versions)
 export(read_cpr_north_data)
 export(read_cpr_south_data)
 export(read_plankton_nets_data)
