### 2025-04-16 526494f
fix: enhance validation for API response
