### 2025-04-16 3a7182c
refactor: streamline version checking and caching logic
