### 2025-04-16 8c3b959
chore: improve error message
