+++ R/load_forcis_test.R
@@ -1,17 +1,61 @@
-#' Load a FORCIS dataset by name
+#' Load a FORCIS dataset
 #'
-#' @param path The directory path to read from
-#' @param version Database version to use, NULL by default
-#' @param name Key identifying the dataset ("net", "pump", etc.)
-#' @param cached Logical, TRUE by default to cache dataset file & metadata
+#' Retrieves and processes a specified FORCIS dataset containing global
+#' foraminifera distribution data. The function handles downloading from
+#' Zenodo, caching and data processing automatically.
 #'
-#' @return A tibble containing the dataset
+#' @param name Key identifying the dataset. Available options:
+#'   \itemize{
+#'     \item \code{"net"} - Plankton Net data
+#'     \item \code{"pump"} - Pump data
+#'     \item \code{"sediment_trap"} - Sediment trap data
+#'     \item \code{"cpr_south"} - CPR (South) data
+#'     \item \code{"cpr_north"} - CPR (North) data
+#'   }
+#' @param version Database version to use. Options:
+#'   \itemize{
+#'     \item \code{NULL} or \code{"latest"} - Use the latest available version
+#'   (default)
+#'     \item Character string - Specific version number (e.g., "01", "02")
+#'   }
+#' @param path The directory path to store or read cached data files.
+#'   If \code{NULL}, uses a default location in the user's cache directory.
+#' @param cached Logical, whether to use cached data files
+#'   (\code{TRUE} by default).
+#'   When \code{FALSE}, files will be downloaded fresh and cleaned up after use.
+#'
+#' @return A tibble containing the requested dataset with appropriate column
+#'   types. Numeric values for species abundance data, categorical values
+#'   for metadata.
+#'
+#' @examples
+#' \dontrun{
+#' # Load the latest version of the plankton net dataset
+#' net_data <- load_forcis_test("net")
+#'
+#' # Load a specific version of the pump dataset without caching
+#' pump_data <- load_forcis_test("pump", version = "02", cached = FALSE)
+#'
+#' # Load sediment trap data with custom cache location
+#' st_data <- load_forcis_test("sediment_trap", path = ".")
+#' }
+#'
+#' @details
+#' The function handles downloading data from Zenodo, verifying file integrity,
+#' and converting columns to appropriate types. When caching is enabled, files
+#' are saved locally to speed up future access to the same dataset.
+#'
+#' The forcis database provides global distribution data for planktonic
+#' foraminifera collected using various methods.
+#'
 #' @noRd
 load_forcis_test <- function(
     name,
     version = NULL,
     path = NULL,
     cached = TRUE) {
+  # TODO: remove '_test' from the function name
+
   ## Check args ----
 
   # Check if name is valid
