### 2025-04-17 c131845
revert: read_* funtions to commit cbaaae6
