+++ R/load_forcis.R
@@ -1,109 +1,262 @@
-# TODO: custom path should be NULL by default (R_user_dir)
-
-#' Load a FORCIS dataset by name
+#' Load a FORCIS dataset
 #'
-#' @param path The directory path to read from
-#' @param version Database version to use (TODO: should be NULL by default)
-#' @param check_for_update Whether to check for database updates
-#'                         (TODO: to be removed)
-#' @param name Key identifying the dataset ("net", "pump", etc.)
-#' @param cached TODO: for later use (cache dataset on local storage)
+#' Retrieves and processes a specified FORCIS dataset containing global
+#' foraminifera distribution data. The function handles downloading from
+#' Zenodo, caching and data processing automatically.
 #'
-#' @return A tibble containing the dataset
-#' @keywords internal
+#' @param name Key identifying the dataset. Available options:
+#'   \itemize{
+#'     \item \code{"net"} - Plankton Net data
+#'     \item \code{"pump"} - Pump data
+#'     \item \code{"sediment_trap"} - Sediment trap data
+#'     \item \code{"cpr_south"} - CPR (South) data
+#'     \item \code{"cpr_north"} - CPR (North) data
+#'   }
+#' @param version Database version to use. Options:
+#'   \itemize{
+#'     \item \code{NULL} or \code{"latest"} - Use the latest available version
+#'   (default)
+#'     \item Character string - Specific version number (e.g., "01", "02")
+#'   }
+#' @param path The directory path to store or read cached data files.
+#'   If \code{NULL}, uses a default location in the user's cache directory.
+#' @param cached Logical, whether to use cached data files
+#'   (\code{TRUE} by default).
+#'   When \code{FALSE}, files will be downloaded fresh and cleaned up after use.
+#'
+#' @return A tibble containing the requested dataset with appropriate column
+#'   types. Numeric values for species abundance data, categorical values
+#'   for metadata.
+#'
+#' @examples
+#' \dontrun{
+#' # Load the latest version of the plankton net dataset
+#' net_data <- load_forcis("net")
+#'
+#' # Load a specific version of the pump dataset without caching
+#' pump_data <- load_forcis("pump", version = "02", cached = FALSE)
+#'
+#' # Load sediment trap data with custom cache location
+#' st_data <- load_forcis("sediment_trap", path = ".")
+#' }
+#'
+#' @details
+#' The function handles downloading data from Zenodo, verifying file integrity,
+#' and converting columns to appropriate types. When caching is enabled, files
+#' are saved locally to speed up future access to the same dataset.
+#'
+#' The forcis database provides global distribution data for planktonic
+#' foraminifera collected using various methods.
+#'
+#' @noRd
 load_forcis <- function(
     name,
-    path = ".",
-    version = options()$"forcis_version",
-    check_for_update = options()$"forcis_check_for_update",
+    version = NULL,
+    path = NULL,
     cached = TRUE) {
+  ## Check args ----
+
   # Check if name is valid
   validate_dataset_name(name = name)
+
+  # Check version
+  if (!is.null(version)) {
+    check_version(version)
+  }
 
   # Extract dataset-specific information
   metadata <- forcis_datasets_info()
   dataset_info <- metadata[[name]]
   display_name <- dataset_info$name
-  file_pattern <- dataset_info$filename_prefix
+  dataset_file_pattern <- dataset_info$filename_prefix
   columns_to_process <- dataset_info$columns
 
-  ## Check args ----
-  check_if_character(path)
-  check_version(version)
+  ## Get metadata from Zenodo ---
 
-  ## Check/set version ----
-  version <- set_version(version, ask = FALSE)
[…]
