+++ R/load_forcis_test.R
@@ -1,264 +1,0 @@
-#' Load a FORCIS dataset
-#'
-#' Retrieves and processes a specified FORCIS dataset containing global
-#' foraminifera distribution data. The function handles downloading from
-#' Zenodo, caching and data processing automatically.
-#'
-#' @param name Key identifying the dataset. Available options:
-#'   \itemize{
-#'     \item \code{"net"} - Plankton Net data
-#'     \item \code{"pump"} - Pump data
-#'     \item \code{"sediment_trap"} - Sediment trap data
-#'     \item \code{"cpr_south"} - CPR (South) data
-#'     \item \code{"cpr_north"} - CPR (North) data
-#'   }
-#' @param version Database version to use. Options:
-#'   \itemize{
-#'     \item \code{NULL} or \code{"latest"} - Use the latest available version
-#'   (default)
-#'     \item Character string - Specific version number (e.g., "01", "02")
-#'   }
-#' @param path The directory path to store or read cached data files.
-#'   If \code{NULL}, uses a default location in the user's cache directory.
-#' @param cached Logical, whether to use cached data files
-#'   (\code{TRUE} by default).
-#'   When \code{FALSE}, files will be downloaded fresh and cleaned up after use.
-#'
-#' @return A tibble containing the requested dataset with appropriate column
-#'   types. Numeric values for species abundance data, categorical values
-#'   for metadata.
-#'
-#' @examples
-#' \dontrun{
-#' # Load the latest version of the plankton net dataset
-#' net_data <- load_forcis_test("net")
-#'
-#' # Load a specific version of the pump dataset without caching
-#' pump_data <- load_forcis_test("pump", version = "02", cached = FALSE)
-#'
-#' # Load sediment trap data with custom cache location
-#' st_data <- load_forcis_test("sediment_trap", path = ".")
-#' }
-#'
-#' @details
-#' The function handles downloading data from Zenodo, verifying file integrity,
-#' and converting columns to appropriate types. When caching is enabled, files
-#' are saved locally to speed up future access to the same dataset.
-#'
-#' The forcis database provides global distribution data for planktonic
-#' foraminifera collected using various methods.
-#'
-#' @noRd
-load_forcis_test <- function(
-    name,
-    version = NULL,
-    path = NULL,
-    cached = TRUE) {
-  # TODO: remove '_test' from the function name
-
-  ## Check args ----
-
-  # Check if name is valid
-  validate_dataset_name(name = name)
-
-  # Check version
-  if (!is.null(version)) {
-    check_version(version)
-  }
-
-  # Extract dataset-specific information
-  metadata <- forcis_datasets_info()
-  dataset_info <- metadata[[name]]
-  display_name <- dataset_info$name
-  dataset_file_pattern <- dataset_info$filename_prefix
-  columns_to_process <- dataset_info$columns
-
-  ## Get metadata from Zenodo ---
-
-  # Initialize zenodo_metadata to NULL
-  zenodo_metadata <- NULL
-
-  # Resolve version for caching setting
-  effective_version <- if (is.null(version) || version == "latest") {
-    get_option("latest_version")
-  } else {
-    version
-  }
-
-  # Calculate meta_cache_path once
-  meta_cache_path <- NULL
-  if (!is.null(effective_version)) {
-    meta_cache_path <- get_meta_cache_path(
-      version = effective_version,
-      path = path
-    )
-  }
-
-  # Handle caching based on settings
-  if (cached && !is.null(meta_cache_path)) {
-    # Load from cache if the file exists
-    if (file.exists(meta_cache_path)) {
-      zenodo_metadata <- readRDS(meta_cache_path)
-    }
-  } else if (!cached && !is.null(meta_cache_path)) {
[…]
