+++ man/load_forcis.Rd
@@ -1,0 +1,63 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/load_forcis.R
+\name{load_forcis}
+\alias{load_forcis}
+\title{Load a FORCIS dataset}
+\usage{
+load_forcis(name, version = NULL, path = NULL, cached = TRUE)
+}
+\arguments{
+\item{name}{Key identifying the dataset. Available options:
+\itemize{
+\item \code{"net"} - Plankton Net data
+\item \code{"pump"} - Pump data
+\item \code{"sediment_trap"} - Sediment trap data
+\item \code{"cpr_south"} - CPR (South) data
+\item \code{"cpr_north"} - CPR (North) data
+}}
+
+\item{version}{Database version to use. Options:
+\itemize{
+\item \code{NULL} or \code{"latest"} - Use the latest available version
+(default)
+\item Character string - Specific version number (e.g., "01", "02")
+}}
+
+\item{path}{The directory path to store or read cached data files.
+If \code{NULL}, uses a default location in the user's cache directory.}
+
+\item{cached}{Logical, whether to use cached data files
+(\code{TRUE} by default).
+When \code{FALSE}, files will be downloaded fresh and cleaned up after use.}
+}
+\value{
+A tibble containing the requested dataset with appropriate column
+types. Numeric values for species abundance data, categorical values
+for metadata.
+}
+\description{
+Retrieves and processes a specified FORCIS dataset containing global
+foraminifera distribution data. The function handles downloading from
+Zenodo, caching and data processing automatically.
+}
+\details{
+The function handles downloading data from Zenodo, verifying file integrity,
+and converting columns to appropriate types. When caching is enabled, files
+are saved locally to speed up future access to the same dataset.
+
+The FORCIS database provides global distribution data for planktonic
+foraminifera collected using various methods.
+}
+\examples{
+\dontrun{
+# Load the latest version of the plankton net dataset
+net_data <- load_forcis("net")
+
+# Load a specific version of the pump dataset without caching
+pump_data <- load_forcis("pump", version = "02", cached = FALSE)
+
+# Load sediment trap data with custom cache location
+st_data <- load_forcis("sediment_trap", path = ".")
+}
+
+}
