### 2025-04-17 5205b1a
refactor(read_*): add a deprecation warning
