+++ R/read_cpr_north_data.R
@@ -5,6 +5,8 @@ read_cpr_north_data <- function(
     path = ".",
     version = options()$"forcis_version",
     check_for_update = options()$"forcis_check_for_update") {
+  .Deprecated("load_forcis")
+
   ## Check args ----
 
   check_if_character(path)
