### 2025-04-17 e94fda6
test(read_*): supress deprecation warnings
