### 2025-04-18 918e178
feat: implement clean_forcis_cache function for managing cached data
