### 2025-04-18 31c80dd
docs: update examples in print_available_versions
