+++ .pre-commit-config.yaml
@@ -2,10 +2,10 @@
 # R specific hooks: https://github.com/lorenzwalthert/precommit
 repos:
   - repo: https://github.com/lorenzwalthert/precommit
-    rev: v0.4.3.9008
+    rev: v0.4.3.9009
     hooks:
-      - id: style-files
-        args: [--style_pkg=styler, --style_fun=tidyverse_style]
+      # - id: style-files
+      #   args: [--style_pkg=styler, --style_fun=tidyverse_style]
       - id: roxygenize
         additional_dependencies:
           - ggplot2
