+++ R/internal_zenodo.R
@@ -2,36 +2,42 @@
 #'
 #' @return Character string with the concept record ID
 #' @noRd
+
 zenodo_conceptrecid <- function() "7390791"
 
 #' Zenodo Record identifier (for v10)
 #'
 #' @return Character string with the record ID
 #' @noRd
+
 zenodo_record_id <- function() "12724286"
 
 #' Zenodo Page Size
 #'
 #' @return Numeric value for page size
 #' @noRd
+
 zenodo_page_size <- function() 999
 
 #' Zenodo API base URL
 #'
 #' @return Character string with the Zenodo API base URL
 #' @noRd
+
 zenodo_api_url <- function() "https://zenodo.org/api"
 
 #' Zenodo Records endpoint URL
 #'
 #' @return Character string with the Zenodo records endpoint URL
 #' @noRd
+
 zenodo_records_endpoint <- function() paste0(zenodo_api_url(), "/records")
 
 #' Zenodo Record Versions endpoint URL
 #'
 #' @return Character string with the Zenodo record versions endpoint URL
 #' @noRd
+
 zenodo_record_versions_url <- function() {
   paste0(zenodo_records_endpoint(), "/", zenodo_record_id(), "/versions")
 }
@@ -40,6 +46,7 @@ zenodo_record_versions_url <- function() {
 #'
 #' @return Character string with the Zenodo latest version endpoint URL
 #' @noRd
+
 zenodo_latest_ver_url <- function() {
   paste0(zenodo_record_versions_url(), "/latest")
 }
@@ -48,6 +55,7 @@ zenodo_latest_ver_url <- function() {
 #'
 #' @return Character string with the error message
 #' @noRd
+
 err_msg_missing_version <- function() {
   paste(
     "Required version missing.",
@@ -56,12 +64,12 @@ err_msg_missing_version <- function() {
 }
 
 
-
 #' Check Zenodo version
 #'
 #' @param version Version string to validate
 #' @return NULL invisibly, raises an error for invalid version
 #' @noRd
+
 check_version <- function(version) {
   if (missing(version)) {
     stop("Argument 'version' is required", call. = FALSE)
@@ -95,6 +103,7 @@ check_version <- function(version) {
 #' @param res API single response from Zenodo
 #' @return Character string with the latest version number
 #' @noRd
+
 get_latest_version <- function(res = NULL) {
   if (is.null(res)) {
     res <- get_metadata(version = "latest")
@@ -110,6 +119,7 @@ get_latest_version <- function(res = NULL) {
 #' @param ask Whether to prompt for upgrading to latest version
 #' @return Character string with the version to use
 #' @noRd
+
 set_version <- function(version, ask = TRUE) {
   check_version(version)
 
@@ -163,6 +173,7 @@ set_version <- function(version, ask = TRUE) {
 #' @param version Version string to save
 #' @return NULL invisibly
 #' @noRd
+
 save_version <- function(version) {
   saved_version <- get_current_version()
 
@@ -179,6 +190,7 @@ save_version <- function(version) {
 #' @param version Version to retrieve metadata for
 #' @return List with JSON response from Zenodo API
 #' @noRd
+
 get_metadata <- function(version = NULL) {
   if (is.null(version)) version <- "latest" # tmp to avoid breaking it
   check_version(version)
@@ -192,7 +204,8 @@ get_metadata <- function(version = NULL) {
       httr2::resp_body_json(http_response)
     },
     error = function(e) {
-      stop("Zenodo API: Failed to retrieve metadata. ",
+      stop(
+        "Zenodo API: Failed to retrieve metadata. ",
         e$message,
         call. = FALSE
       )
@@ -205,8 +218,10 @@ get_metadata <- function(version = NULL) {
 #' @param version Version to build request for
 #' @return httr2 request object
 #' @noRd
+
[…]
