### 2025-04-18 bcb3ae8
test: add tests for print_available_versions()
