+++ DESCRIPTION
@@ -66,5 +66,6 @@ Suggests:
     testthat (>= 3.0.0),
     vdiffr,
     withr,
-    lintr
+    lintr,
+    mockery
 Config/testthat/edition: 3
