### 2025-04-18 5ea98b1
test: add tests for clean_forcis_cache
