+++ tests/testthat/test-clean_forcis_cache.R
@@ -1,0 +1,192 @@
+test_that("clean_forcis_cache removes entire cache when version is NULL", {
+  # Create a temporary root directory
+  root_dir <- tempfile("forcis_test_root")
+  dir.create(root_dir, recursive = TRUE)
+
+  # Create the expected directory structure
+  data_dir <- file.path(root_dir, "forcis", "data")
+  dir.create(data_dir, recursive = TRUE)
+
+  # Create version directories
+  v1_dir <- file.path(data_dir, "01")
+  v2_dir <- file.path(data_dir, "02")
+  dir.create(v1_dir)
+  dir.create(v2_dir)
+
+  # Add some test files
+  writeLines("test", file.path(v1_dir, "data.csv"))
+  writeLines("test", file.path(v2_dir, "metadata.json"))
+
+  # Capture messages
+  messages <- capture_messages({
+    result <- clean_forcis_cache(path = root_dir)
+  })
+
+  # Check that function returns TRUE
+  expect_true(result)
+
+  # Check that correct message was output
+  expect_match(messages, paste0("Cleaned all cached data: ", data_dir))
+
+  # Check that directory was removed
+  expect_false(dir.exists(data_dir))
+})
+
+test_that("clean_forcis_cache handles non-existent cache directory", {
+  # Use a path that definitely doesn't exist
+  non_existent_dir <- tempfile("forcis_nonexistent")
+
+  # Capture messages
+  messages <- capture_messages({
+    result <- clean_forcis_cache(path = non_existent_dir)
+  })
+
+  # Check that function returns TRUE
+  expect_true(result)
+
+  # Check that correct message was output
+  expect_match(messages, "Cache directory does not exist, nothing to clean")
+})
+
+test_that("clean_forcis_cache removes specific version when version is provided", {
+  # Create a temporary root directory
+  root_dir <- tempfile("forcis_test_root")
+
+  # Create the expected directory structure
+  data_dir <- file.path(root_dir, "forcis", "data")
+  dir.create(data_dir, recursive = TRUE)
+
+  # Create version directories
+  v1_dir <- file.path(data_dir, "01")
+  v2_dir <- file.path(data_dir, "02")
+  dir.create(v1_dir)
+  dir.create(v2_dir)
+
+  # Add some test files
+  writeLines("test", file.path(v1_dir, "data.csv"))
+  writeLines("test", file.path(v2_dir, "metadata.json"))
+
+  # Capture messages
+  messages <- capture_messages({
+    result <- clean_forcis_cache(version = "01", path = root_dir)
+  })
+
+  # Check that function returns TRUE
+  expect_true(result)
+
+  # Check that correct message was output
+  expect_match(messages, "Cleaned cache for version 01")
+
+  # Check that only version 01 directory was removed
+  expect_false(dir.exists(v1_dir))
+  expect_true(dir.exists(v2_dir))
+})
+
+test_that("clean_forcis_cache handles non-existent version", {
+  # Create a temporary root directory
+  root_dir <- tempfile("forcis_test_root")
+
+  # Create the expected directory structure
+  data_dir <- file.path(root_dir, "forcis", "data")
+  dir.create(data_dir, recursive = TRUE)
+
+  # Create version directory (but not version 99)
+  v1_dir <- file.path(data_dir, "01")
+  dir.create(v1_dir)
+
+  # Capture warnings
+  warnings <- capture_warnings({
+    result <- clean_forcis_cache(version = "99", path = root_dir)
+  })
+
+  # Check that function returns FALSE
+  expect_false(result)
+
+  # Check that correct warning was output
+  expect_match(warnings, "Version 99 is not cached")
+})
+
[…]
