### 2025-04-21 c355143
chore: use NULL as default inget_version_metadata()
