+++ R/get_version_metadata.R
@@ -6,7 +6,7 @@
 #'
 #' @param version a `character` of length 1. The label of the version. Use
 #'   [get_available_versions()] to list available versions.
-#'   If `latest` (default) the latest version is used.
+#'   If `NULL` (default) the latest version is used.
 #'
 #' @return A `list` with all information about the version, including: `title`,
 #' `doi`, `publication_date`, `description`, `access_right`, `creators`,
@@ -21,14 +21,14 @@
 #' # Get information for the latest version of the FORCIS database ----
 #' get_version_metadata()
 
-get_version_metadata <- function(version = "latest") {
+get_version_metadata <- function(version = NULL) {
   check_version(version)
 
   # Retrieve metadata
   res <- get_metadata(version)
 
   # Extract versions and determine position
-  if (version == "latest") {
+  if (is.null(version) || version == "latest") {
     versions <- extract_single_version(res)
     pos <- 1 # Since it's a single item
   } else {
