+++ man/get_version_metadata.Rd
@@ -4,12 +4,12 @@
 \alias{get_version_metadata}
 \title{Print information of a specific version of the FORCIS database}
 \usage{
-get_version_metadata(version = "latest")
+get_version_metadata(version = NULL)
 }
 \arguments{
 \item{version}{a \code{character} of length 1. The label of the version. Use
 \code{\link[=get_available_versions]{get_available_versions()}} to list available versions.
-If \code{latest} (default) the latest version is used.}
+If \code{NULL} (default) the latest version is used.}
 }
 \value{
 A \code{list} with all information about the version, including: \code{title},
