### 2025-04-21 8edbad8
Fix: get_version_metadata checks
