### 2025-04-21 f602c51
test(clean_forcis_cache): replace hardcoded paths with get_data_dir
