### 2025-04-21 8f09b39
fix(load_forcis): correct version handling for caching logic pkg options could be tampered, so force api check for lastest version.
