### 2025-04-22 63542f9
test(load_forcis): add a case for latest version
