### 2025-04-22 28f6690
test(load_forcis): refactor the test
