### 2025-04-22 21decc2
test(load_forcis): add a case for a specific version
