+++ tests/testthat/test-load_forcis_data.R
@@ -9,6 +9,21 @@ forcis_datasets <- forcis_datasets_info()
 
 ## Helper functions ---
 
+# mock downloading a file by creating a dummy file
+download_file_mock <- function(url, file, path, overwrite = FALSE, ...) {
+  dir.create(path, recursive = TRUE, showWarnings = FALSE)
+
+  dest_path <- file.path(path, file)
+  log_message("Writing mock data to:", dest_path)
+  writeLines(MOCK_CSV_DATA, dest_path)
+  invisible(NULL)
+}
+
+# always return TRUE for checksum verification
+verify_file_checksum_mock <- function(file_path, expected_checksum) {
+  return(TRUE)
+}
+
 # set of data checks for a dataset
 verify_result <- function(
   result,
@@ -66,20 +81,6 @@ test_that("Test load_forcis (latest dataset version) for success", {
   # Mock the get_metadata function to return our controlled metadata
   get_metadata_mock <- function(...) {
     return(mock_metadata_env$metadata)
-  }
-
-  download_file_mock <- function(url, file, path, overwrite = FALSE, ...) {
-    dir.create(path, recursive = TRUE, showWarnings = FALSE)
-
-    dest_path <- file.path(path, file)
-    log_message("Writing mock data to:", dest_path)
-    writeLines(MOCK_CSV_DATA, dest_path)
-    invisible(NULL)
-  }
-
-  # Always return TRUE for checksum verification
-  verify_file_checksum_mock <- function(file_path, expected_checksum) {
-    return(TRUE)
   }
 
   ## Load a dataset (cached = TRUE) ---
@@ -150,4 +151,115 @@ test_that("Test load_forcis (latest dataset version) for success", {
   )
 })
 
-# TODO: add more test cases (specific versions & datasets)
+# Test loading a dataset (version 08)
+test_that("Test load_forcis (specific dataset version) for success", {
+  root_dir <- tempfile("load_forcis_root")
+
+  log_message("load_forcis_root :", root_dir)
+
+  # Mock the get_metadata function
+  mock_metadata_env <- new.env()
+
+  with_mock_dir(
+    test_path("mockdata", "v08"),
+    {
+      mock_metadata_env$metadata <- get_metadata("08")
+    },
+    simplify = FALSE
+  )
+
+  # Mock the get_metadata function to return our controlled metadata
+  get_metadata_mock <- function(...) {
+    return(mock_metadata_env$metadata)
+  }
+
+  ## Load a dataset (cached = TRUE) ---
+
+  with_mocked_bindings(
+    get_metadata = get_metadata_mock,
+    download_file = download_file_mock,
+    verify_file_checksum = verify_file_checksum_mock,
+    {
+      messages <- capture_messages({
+        result <- load_forcis("net", version = "08", path = root_dir)
+      })
+      messages_cache <- capture_messages({
+        result_cache <- load_forcis("net", version = "08", path = root_dir)
+      })
+    }
+  )
+
+  # dataset filename prefix
+  dataset_net_prefix <- forcis_datasets$filename_prefixes()["net"]
+
+  # Expect a message for a missing file
+  expect_match(
+    messages,
+    paste0("File ", dataset_net_prefix, ".*\\.csv is missing, downloading...")
+  )
+
+  # Verify the result
+  verify_result(result, 2, 8, c(10, 8))
+
+  # Expect no messages for the 2nd call
+  expect_length(messages_cache, 0)
+
+  # 2nd call (cache)
+  verify_result(result_cache, 2, 8, c(10, 8))
+
+  # Check cache
+  verify_dataset_cache(
+    root_dir,
+    dataset_net_prefix,
+    expect_files_exist = TRUE
+  )
+
[…]
