### 2025-04-23 7217f5e
test(load_forcis): cover all needed cases
