### 2025-04-23 19c0c56
rOpenSci review: Add a statement of need (#84)
