+++ README.Rmd
@@ -7,9 +7,9 @@ output: github_document
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
-  collapse  = TRUE,
-  comment   = "#>",
-  fig.path  = "man/figures/",
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/",
   out.width = "100%"
 )
 ```
@@ -37,7 +37,6 @@ forcis <img src="man/figures/logo.png" height="120px" align="right" style="float
 
 <p align="left">
   • <a href="#overview">Overview</a><br>
-  • <a href="#features">Features</a><br>
   • <a href="#installation">Installation</a><br>
   • <a href="#get-started">Get started</a><br>
   • <a href="#long-form-documentations">Long-form documentations</a><br>
@@ -53,11 +52,7 @@ forcis <img src="man/figures/logo.png" height="120px" align="right" style="float
 
 The goal of the R package `forcis` is to provide an interface to the [FORCIS database](https://zenodo.org/doi/10.5281/zenodo.7390791) on global foraminifera distribution (Chaabane _et al._ 2023). This database includes data on living planktonic foraminifera diversity and distribution in the global oceans from 1910 until 2018 collected using plankton tows, continuous plankton recorder, sediment traps and plankton pump from the global ocean.
 
-
-
-## Features
-
-The main purpose of the `forcis` package is to download, select, filter, homogenize, and visualize FORCIS data. The [Functions](https://frbcesab.github.io/forcis/reference/index.html) page provides a full description of the features available in the package.
+This package has been developed for researchers interested in working with the FORCIS database, even without advanced R skills. It provides basic functions to facilitate the handling of this large database, including functions to download, select, filter, homogenize, and visualize the data. It also enables users to explore the spatial distribution and temporal evolution of planktonic foraminifera. The [Functions](https://frbcesab.github.io/forcis/reference/index.html) page provides a full description of the features available in the package.
 
 
 ## Installation
