### 2025-04-23 3ff0618
factor: refactor read_*_data() functions
