+++ R/internals.R
@@ -778,3 +778,79 @@ add_data_type <- function(data, type) {
 
   data
 }
+
+
+#' Read a FORCIS datasheet
+#'
+#' @noRd
+
+read_data_ <- function(
+  check_file_fun,
+  data_msg,
+  path = ".",
+  version = options()$"forcis_version",
+  check_for_update = options()$"forcis_check_for_update"
+) {
+  ## Check args ----
+
+  check_if_character(path)
+  check_version(version)
+
+  ## Check/set version ----
+  version <- set_version(version, ask = FALSE)
+
+  ## Check local database ----
+  path <- file.path(path, "forcis-db", paste0("version-", version))
+
+  if (!dir.exists(path)) {
+    stop(
+      "The directory '",
+      path,
+      "' does not exist. Please check the ",
+      "argument 'path' or use the function 'download_forcis_db()'.",
+      call. = FALSE
+    )
+  }
+
+  ## Check file ----
+  file_name <- list.files(path, pattern = check_file_fun())
+
+  if (!length(file_name)) {
+    stop(
+      sprintf(
+        "The %s dataset does not exist. Please use the function ",
+        data_msg
+      ),
+      "'download_forcis_db()'.",
+      call. = FALSE
+    )
+  }
+
+  ## Check for update ----
+
+  if (is.null(check_for_update)) {
+    check_for_update <- TRUE
+  }
+
+  if (check_for_update) {
+    if (version != get_latest_version()) {
+      message(
+        "A newer version of the FORCIS database is available. Use ",
+        "'download_forcis_db(version = NULL)' to download it."
+      )
+    }
+  }
+
+  ## Read data ----
+
+  file_name <- list.files(path, pattern = check_file_fun())
+
+  data <- vroom::vroom(
+    file.path(path, file_name),
+    delim = ";",
+    altrep = FALSE,
+    show_col_types = FALSE
+  )
+
+  as.data.frame(data)
+}
