### 2025-04-23 6fe404b
rOpenSci review: Refactor read_*_data() functions (#85)
