### 2025-04-23 8f38299
chore: split internals.R in separate files
