+++ R/utils-filenames.R
@@ -1,0 +1,33 @@
+#' Plankton nets file name
+#'
+#' @noRd
+
+plankton_net_filename <- function() "FORCIS_net_"
+
+
+#' Pumps file name
+#'
+#' @noRd
+
+pump_filename <- function() "FORCIS_pump_"
+
+
+#' CPR North file name
+#'
+#' @noRd
+
+cpr_north_filename <- function() "FORCIS_cpr_north_"
+
+
+#' CPR South file name
+#'
+#' @noRd
+
+cpr_south_filename <- function() "FORCIS_cpr_south_"
+
+
+#' Sediment Traps file name
+#'
+#' @noRd
+
+sediment_trap_filename <- function() "FORCIS_trap_"

[…]
