### 2025-04-23 eab679e
chore: remove unused function
