### 2025-04-23 672892a
Reorder functions in files (#86)
* chore: split internals.R in separate files

* chore: remove unused function
