+++ R/internals.R
@@ -1,856 +1,0 @@
-#' Download a csv file
-#'
-#' @noRd
-
-download_file <- function(url, path, file, overwrite = FALSE, timeout = 60) {
-  url <- utils::URLencode(url)
-  file <- gsub("\\s", "_", file)
-
-  check_if_path_exists(path)
-
-  ## Check if the file already exists ----
-
-  destination <- file.path(path, file)
-
-  if (!overwrite && file.exists(destination)) {
-    message(
-      "The file '",
-      file,
-      "' already exists. If you want to download ",
-      "again this file please use the argument 'overwrite'."
-    )
-    return(invisible(NULL))
-  }
-
-  ## Download the file if 'overwrite' is TRUE or it doesn't exist ----
-
-  # change timeout for large file and slow connection
-  user_opts <- options()
-  on.exit(options(user_opts))
-  options(timeout = max(timeout, getOption("timeout")))
-
-  tryCatch(
-    {
-      utils::download.file(url = url, destfile = destination, mode = "wb")
-
-      message("The file '", file, "' has been successfully downloaded")
-    },
-    error = function(e) {
-      message("Download error: ", e$message)
-
-      if (file.exists(destination)) {
-        file.remove(destination)
-
-        message("Temporary file deleted")
-      }
-    }
-  )
-
-  invisible(NULL)
-}
-
-
-#' Plankton nets file name
-#'
-#' @noRd
-
-plankton_net_filename <- function() "FORCIS_net_"
-
-
-#' Pumps file name
-#'
-#' @noRd
-
-pump_filename <- function() "FORCIS_pump_"
-
-
-#' CPR North file name
-#'
-#' @noRd
-
-cpr_north_filename <- function() "FORCIS_cpr_north_"
-
-
-#' CPR South file name
-#'
-#' @noRd
-
-cpr_south_filename <- function() "FORCIS_cpr_south_"
-
-
-#' Sediment Traps file name
-#'
-#' @noRd
-
-sediment_trap_filename <- function() "FORCIS_trap_"
-
-
-#' Vector of device types available in FORCIS database
-#'
-#' @noRd
-
-data_types <- function()
-  c("Net", "Pump", "Sediment trap", "CPR South", "CPR North")
-
-
-#' Species names and taxonomy name
-#'
-#' @noRd
-
-species_list <- function() {
-  data.frame(
-    "taxon" = c(
-      "b_digitata_LT",
-      "b_pumilio_LT",
-      "b_variabilis_LT",
-      "c_nitida_LT",
-      "d_anfracta_LT",
-      "g_adamsi_LT",
-      "g_bulloides_LT",
-      "g_calida_LT",
-      "g_cavernula_LT",
-      "g_conglobatus_LT",
-      "g_conglomerata_LT",
-      "g_crassaformis_LT",
-      "g_falconensis_LT",
-      "g_glutinata_LT",
-      "g_hexagona_LT",
-      "g_hirsuta_LT",
-      "g_inflata_LT",
-      "g_cultrata_LT",
-      "g_minuta_LT",
-      "g_ruber_any_LT",
-      "g_rubescens_LT",
-      "g_scitula_LT",
-      "g_siphonifera_LT",
-      "g_tenellus_LT",
-      "g_theyeri_LT",
-      "g_truncatulinoides_LT",
-      "g_tumida_LT",
-      "g_ungulata_LT",
-      "g_uvula_LT",
-      "n_vivans_LT",
-      "h_digitata_LT",
-      "h_pelagica_LT",
-      "n_dutertrei_LT",
-      "n_incompta_LT",
-      "n_pachyderma_any_LT",
-      "o_riedeli_LT",
-      "o_universa_LT",
-      "p_obliquiloculata_LT",
[…]
