+++ R/utils.R
@@ -1,0 +1,184 @@
+#' Download a csv file
+#'
+#' @noRd
+
+download_file <- function(url, path, file, overwrite = FALSE, timeout = 60) {
+  url <- utils::URLencode(url)
+  file <- gsub("\\s", "_", file)
+
+  check_if_path_exists(path)
+
+  ## Check if the file already exists ----
+
+  destination <- file.path(path, file)
+
+  if (!overwrite && file.exists(destination)) {
+    message(
+      "The file '",
+      file,
+      "' already exists. If you want to download ",
+      "again this file please use the argument 'overwrite'."
+    )
+    return(invisible(NULL))
+  }
+
+  ## Download the file if 'overwrite' is TRUE or it doesn't exist ----
+
+  # change timeout for large file and slow connection
+  user_opts <- options()
+  on.exit(options(user_opts))
+  options(timeout = max(timeout, getOption("timeout")))
+
+  tryCatch(
+    {
+      utils::download.file(url = url, destfile = destination, mode = "wb")
+
+      message("The file '", file, "' has been successfully downloaded")
+    },
+    error = function(e) {
+      message("Download error: ", e$message)
+
+      if (file.exists(destination)) {
+        file.remove(destination)
+
+        message("Temporary file deleted")
+      }
+    }
+  )
+
+  invisible(NULL)
+}
+
+
+#' Date format used in raw data
+#'
+#' @noRd
+
+date_format <- function() "%d/%m/%Y"
+
+
+#' Retrieve data type
+#'
+#' @noRd
+
+get_data_type <- function(data) {
+  check_if_df(data)
+  check_field_in_data(data, "data_type")
+
+  data_type <- unique(data$"data_type")
+
+  if (length(data_type) != 1) {
+    stop(
+      "The column 'data_type' cannot contain different values",
+      call. = FALSE
+    )
+  }
+
+  data_type
+}
+
+
+#' Add a column 'data_type' in data.frame (if required)
+#'
+#' @noRd
+
+add_data_type <- function(data, type) {
+  check_if_df(data)
+  check_if_character(type)
+
+  if ("data_type" %in% colnames(data)) {
+    data$"data_type" <- type
+  } else {
+    data <- data.frame("data_type" = type, data)
+  }
+
+  data
+}
+
+
+#' Read a FORCIS datasheet
+#'
+#' @noRd
+
+read_data_ <- function(
+  check_file_fun,
+  data_msg,
+  path = ".",
+  version = options()$"forcis_version",
+  check_for_update = options()$"forcis_check_for_update"
+) {
+  ## Check args ----
+
+  check_if_character(path)
+  check_version(version)
+
+  ## Check/set version ----
+  version <- set_version(version, ask = FALSE)
+
+  ## Check local database ----
+  path <- file.path(path, "forcis-db", paste0("version-", version))
+
+  if (!dir.exists(path)) {
+    stop(
+      "The directory '",
+      path,
+      "' does not exist. Please check the ",
+      "argument 'path' or use the function 'download_forcis_db()'.",
+      call. = FALSE
+    )
+  }
+
+  ## Check file ----
+  file_name <- list.files(path, pattern = check_file_fun())
+
+  if (!length(file_name)) {
+    stop(
+      sprintf(
+        "The %s dataset does not exist. Please use the function ",
+        data_msg
+      ),
+      "'download_forcis_db()'.",
+      call. = FALSE
+    )
+  }
+
+  ## Check for update ----
+
+  if (is.null(check_for_update)) {
+    check_for_update <- TRUE
+  }
+
+  if (check_for_update) {
[…]
