### 2025-04-23 17fd4c3
feat: delete and adapt functions linked to .forcis
