### 2025-04-23 478df59
test: adapt tests
