### 2025-04-23 097a811
rOpenSci review: Remove hidden dotfile (#87)
* feat: delete and adapt functions linked to .forcis

* test: adapt tests

* doc: remove .forcis explanation in vignette
