+++ NAMESPACE
@@ -14,7 +14,6 @@ export(filter_by_species)
 export(filter_by_year)
 export(geom_basemap)
 export(get_available_versions)
-export(get_current_version)
 export(get_ocean_names)
 export(get_required_columns)
 export(get_species_names)
