### 2025-04-23 eaf48ff
rOpenSci review: Fix unit test failure (#88)
