### 2025-04-23 718b4cb
doc: update NEWS (reviewer 1)
