+++ NEWS.md
@@ -1,10 +1,16 @@
 # forcis 0.1.0.9000
 
-Answers to editor and reviewer comments.
+Answers to editor and reviewer comments:
 
+* Refactor code of `read_*_data()` functions
+* Recommend to use [`air`](https://github.com/posit-dev/air) to automatically format code ([#80](https://github.com/FRBCesab/forcis/pull/80))
+* Homogeneize function outputs (`tibble`) ([#82](https://github.com/FRBCesab/forcis/pull/82))
+* Use the same prefix `forcis_` in global options ([#81](https://github.com/FRBCesab/forcis/pull/81))
+* Remove the usage of the hidden dotfile `.forcis` ([#87](https://github.com/FRBCesab/forcis/pull/87))
+* Add a statement of need ([#84](https://github.com/FRBCesab/forcis/pull/84))
 * Improve documentation ([#64](https://github.com/FRBCesab/forcis/issues/64), [#65](https://github.com/FRBCesab/forcis/issues/65), [#66](https://github.com/FRBCesab/forcis/issues/66))
 * Add funder in `DESCRIPTION` ([#69](https://github.com/FRBCesab/forcis/issues/69))
-* Improve unit tests ([#67](https://github.com/FRBCesab/forcis/issues/67), [#68](https://github.com/FRBCesab/forcis/issues/68))
+* Improve unit tests ([#67](https://github.com/FRBCesab/forcis/issues/67), [#68](https://github.com/FRBCesab/forcis/issues/68), [#79](https://github.com/FRBCesab/forcis/pull/79), [#88](https://github.com/FRBCesab/forcis/pull/88))
 
 
 
