### 2025-04-23 f9b33c5
doc: update NEWS (reviewer 1)
