### 2025-04-23 #660 @ahasverus
Dear @sckott,

Thanks for your review (and sorry for my late reply).

Here are the answers to your comments:

- Unit test failure in `download_forcis_db()`

> Fix in [PR #88](https://github.com/FRBCesab/forcis/pull/88). All unit tests pass on GitHub Actions (and locally).

- Warnings in unit tests (`tidyselect` and `.data$`)

> Fix in [PR #79](https://github.com/FRBCesab/forcis/pull/79). I refactored the code of `compute_*()` and `plot_record_by_*()` functions using the base R syntax (and by removing dependency to `dplyr` and `tidyselect` packages).

- Statement of need

> I have added a paragraph in the [`README`](https://github.com/FRBCesab/forcis/blob/main/README.md) and the [Get started](https://frbcesab.github.io/forcis/articles/forcis.html) vignette in [PR #84](https://github.com/FRBCesab/forcis/pull/84). I hope this change help informing visitor about the targetted audience of the package.

- The hidden dotfile `.forcis`

> I decided to remove the usage this `.forcis` file. I don't know why I decided to do this in the first place. Now, by default, the package uses the latest version of the database on Zenodo, unless the user wants to use a specific version. He/she will have to use the argument `version` of the `read_*_data()` and `download_forcis_db()` functions or the global option `forcis_version`. See [PR #87](https://github.com/FRBCesab/forcis/pull/87)

- Prefix in global options

> Thanks for this suggestion. Fix in [PR #81](https://github.com/FRBCesab/forcis/pull/81). The two global options start now with `forcis_`.

- Check parameters with no default value

> As the audience of this package is beginners in R, we would like to keep explicit error messages. Is that OK?

- Heterogeous function outputs (`data.frame` and `tibble`)

> Fix in [PR #82](https://github.com/FRBCesab/forcis/pull/82). All exported functions now return `tibble`.

- Style guide in the CONTRIBUTING file

> Thanks for the suggestion. I decided to use `air` (as I switched to Positron IDE) to automatically format code. I have added a `air.toml` file and a few lines in the [`CONTRIBUTING.md`](https://github.com/FRBCesab/forcis/blob/main/CONTRIBUTING.md#style-guide) file. See [PR #80](https://github.com/FRBCesab/forcis/pull/80).

- About the `rappdirs` package

> Thanks for the discovery. I will definitively use this package for my projects. But for this package, as the audience is beginners in R, I guess they want to easily have access to the csv files. We would like to let them choose the directory to save FORCIS database. What do you think?

- Code refactoring in `read_*_data()` functions

> Thanks for the suggestion. I have implemented your code in [PR #85](https://github.com/FRBCesab/forcis/pull/85). Note that I have also included the `vroom::vroom()` section in the helper function that now returns a `data.frame` (and not the path).

I have listed all these changes in the [`NEWS.md`](https://github.com/FRBCesab/forcis/blob/main/NEWS.md).

Thanks again for your review.
Nicolas
