### 2025-04-23 #660 @sckott
@ahasverus

> As the audience of this package is beginners in R, we would like to keep explicit error messages. Is that OK?

yeah, sound good

> We would like to let them choose the directory to save FORCIS database. What do you think?

that sounds fine to not use the pkg. just make sure to not write files to disk in your examples, tests, vignettes, etc. - anything that runs during R CMD CHECK - which cran checks will complain about - write them to temporary directories/files
