### 2025-04-24 FRBCesab/forcis#92 @ahasverus
rOpenSci review: Add a figure on package main features
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2759474577):

> I think a small example of the package functionality on the README page or the Get Started page would be a great addition. I had to click around a lot before I understood what the package use would look like, even after running the code on the get started page.

Here I propose to add a figure showing main features of the package in the `README`.
