### 2025-04-24 FRBCesab/forcis#93 @ahasverus
rOpenSci review: Add a use case
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2759474577):

> I was still pretty unsure what the purpose and final result of using this package was at the end of this page. I recommend including a small example here! Even as simple as a head() and small description of what I am looking at would provide a bit more context to users.

I propose to add a use case in the `Get started` vignette showing an example of the package workflow.
