### 2025-04-24 FRBCesab/forcis#94 @ahasverus
rOpenSci review: Improve description of taxonomies
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2759474577):

> Linking to a short vignette describing the taxonomy codes (similar in scope to the Sampling devices vignette) might be less intimidating than linking to the data paper... or at least providing and indication of which section in the data paper to read about this would be helpful! It was not immediately clear from the paper headings where to find this information.

Add a short description of taxonomies in [`select_taxonomy()`](https://frbcesab.github.io/forcis/reference/select_taxonomy.html) function header and in the vignette [Select, reshape, and filter data](https://frbcesab.github.io/forcis/articles/select-and-filter-data.html#select-a-taxonomy)
