### 2025-04-24 FRBCesab/forcis#95 @ahasverus
rOpenSci review: Improve Select, reshape, and filter data vignette
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2759474577):

> In the selecting columns section, it was unclear to me if you were showing two different selection methods, or a multistep selection method... are you showing a filter of VT data and then forcis columns? perhaps naming the objects differently and/or expanding the commented lines would make this clearer.

Better explain the workflow.
