### 2025-04-24 FRBCesab/forcis#96 @ahasverus
rOpenSci review: Homogeneize object names in all vignettes
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2759474577):

> Also, the same selection code was used on the Get Started page but the object was given a different name... could consider keeping this consistent.
