### 2025-04-24 FRBCesab/forcis#97 @ahasverus
rOpenSci review: Add plots in README
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2759474577):

> These vignettes were the first time I felt I understood the point of the package/some potential uses. Having a small visualization example like these on the README or Get Started to show a start to finish workflow to a sensible end result would be helpful even if its redundant with the more detailed vignettes.

Let's add a map and a barplot in the `README`.
