### 2025-04-24 FRBCesab/forcis#98 @ahasverus
rOpenSci review: Remove library(forcis) in example
The following point has been raised in the [review](https://github.com/ropensci/software-review/issues/660#issuecomment-2759474577):

> Looks great! Some of your examples include library("forcis") before reading example data and some don't...perhaps commit one way or the other. I think it's usually pretty standard to exclude these for the package itself, but should be done you're attaching a dependency (like for ggplot2 in [this example](https://github.com/FRBCesab/forcis/blob/9926fdc8557e7f41a79664d8e0efcc18aadb9b4c/R/geom_basemap.R#L16))

Need to remove `library("forcis")` in example
