### 2025-04-24 6147118
doc: remove library(forcis) from function examples
