+++ R/select_forcis_columns.R
@@ -17,9 +17,6 @@
 #' @return A `tibble`.
 #'
 #' @examples
-#' # Attach the package ----
-#' library("forcis")
-#'
 #' # Import example dataset ----
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 #'                          package = "forcis")
