+++ R/read_plankton_nets_data.R
@@ -28,9 +28,6 @@
 #'
 #' @examples
 #' \dontrun{
-#' # Attach the package ----
-#' library("forcis")
-#'
 #' # Folder in which the database will be saved ----
 #' path_to_save_db <- "data"
 #'
