+++ man/filter_by_year.Rd
@@ -19,9 +19,6 @@ A \code{tibble} containing a subset of \code{data} for the desired years.
 Filters FORCIS data by year of sampling.
 }
 \examples{
-# Attach the package ----
-library("forcis")
-
 # Import example dataset ----
 file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
