+++ man/ggmap_data.Rd
@@ -20,9 +20,6 @@ A \code{ggplot} object.
 Maps the spatial distribution of FORCIS data.
 }
 \examples{
-# Attach the package ----
-library("forcis")
-
 # Import example dataset ----
 file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
