### 2025-04-24 17c19b9
doc: dontrun examples querying zenodo api
