+++ R/get_version_metadata.R
@@ -15,8 +15,10 @@
 #' @export
 #'
 #' @examples
+#' \dontrun{
 #' # Get information for the latest version of the FORCIS database ----
 #' get_version_metadata()
+#' }
 
 get_version_metadata <- function(version = NULL) {
   ## Check arguments ----
