+++ R/get_available_versions.R
@@ -12,8 +12,10 @@
 #' @export
 #'
 #' @examples
+#' \dontrun{
 #' # Versions of the FORCIS database ----
 #' get_available_versions()
+#' }
 
 get_available_versions <- function() {
   ## Retrieve information ----
