+++ man/get_version_metadata.Rd
@@ -21,6 +21,8 @@ Prints information of a specific version of the FORCIS database by querying
 the Zenodo API (\url{https://developers.zenodo.org}).
 }
 \examples{
+\dontrun{
 # Get information for the latest version of the FORCIS database ----
 get_version_metadata()
+}
 }
