+++ man/get_available_versions.Rd
@@ -19,6 +19,8 @@ Gets all available versions of the FORCIS database by querying the Zenodo API
 (\url{https://developers.zenodo.org}).
 }
 \examples{
+\dontrun{
 # Versions of the FORCIS database ----
 get_available_versions()
+}
 }
