### 2025-04-24 1f6cf9a
doc: run function examples
