### 2025-04-24 241d80c
tests: refactor mocking data for offline mode
