+++ tests/testthat/get_version_metadata/zenodo.org/api/records-f5a5db.R
@@ -1,16 +1,0 @@
-structure(list(method = "GET", url = "https://zenodo.org/api/records?q=conceptrecid:7390791&all_versions=true", 
-    status_code = 200L, headers = structure(list(server = "nginx", 
-        date = "Wed, 23 Apr 2025 15:23:42 GMT", `content-type` = "application/json", 
-        `transfer-encoding` = "chunked", vary = "Accept-Encoding", 
-        `access-control-allow-origin` = "*", `access-control-expose-headers` = "Content-Type, ETag, Link, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset", 
-        `x-ratelimit-limit` = "133", `x-ratelimit-remaining` = "33", 
-        `x-ratelimit-reset` = "1745421826", `retry-after` = "3", 
-        `permissions-policy` = "interest-cohort=()", `x-frame-options` = "sameorigin", 
-        `x-xss-protection` = "1; mode=block", `x-content-type-options` = "nosniff", 
-        `content-security-policy` = "default-src 'self' fonts.googleapis.com *.gstatic.com data: 'unsafe-inline' 'unsafe-eval' blob: zenodo-broker.web.cern.ch zenodo-broker-qa.web.cern.ch maxcdn.bootstrapcdn.com cdnjs.cloudflare.com ajax.googleapis.com webanalytics.web.cern.ch", 
-        `strict-transport-security` = "max-age=31556926; includeSubDomains", 
-        `referrer-policy` = "strict-origin-when-cross-origin", 
-        `strict-transport-security` = "max-age=15768000", `x-request-id` = "a0a006a37d6a0340a49454209a0efb2e", 
-        `content-encoding` = "gzip"), redact = character(0), class = "httr2_headers"), 
[…]
