### 2025-04-24 17da126
doc: remove creation of data_type column in function examples
