+++ R/ggmap_data.R
@@ -20,9 +20,6 @@
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #'
-#' # Add 'data_type' column ----
-#' net_data$"data_type" <- "Net"
-#'
 #' # Map data (default) ----
 #' ggmap_data(net_data)
 #'
