+++ R/filter_by_bbox.R
@@ -25,9 +25,6 @@
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #'
-#' # Add 'data_type' column ----
-#' net_data$"data_type" <- "Net"
-#'
 #' # Dimensions of the data.frame ----
 #' dim(net_data)
 #'
