+++ R/data_to_sf.R
@@ -22,9 +22,6 @@
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
 #'
-#' # Add 'data_type' column ----
-#' net_data$"data_type" <- "Net"
-#'
 #' # Dimensions of the data.frame ----
 #' dim(net_data)
 #'
