+++ man/filter_by_polygon.Rd
@@ -26,9 +26,6 @@ file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 
 net_data <- read.table(file_name, dec = ".", sep = ";")
 
-# Add 'data_type' column ----
-net_data$"data_type" <- "Net"
-
 # Dimensions of the data.frame ----
 dim(net_data)
 
