### 2025-04-24 729bdf5
rOpenSci review: Clean examples (#99)
* doc: remove library(forcis) from function examples

* doc: dontrun examples querying zenodo api

* doc: run function examples

* doc: remove creation of data_type column in function examples
