+++ R/select_taxonomy.R
@@ -16,17 +16,11 @@
 #' @return A `tibble`.
 #'
 #' @examples
-#' # Attach the package ----
-#' library("forcis")
-#'
 #' # Import example dataset ----
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 #'                          package = "forcis")
 #'
 #' net_data <- read.table(file_name, dec = ".", sep = ";")
-#'
-#' # Add 'data_type' column ----
-#' net_data$"data_type" <- "Net"
 #'
 #' # Dimensions of the data.frame ----
 #' dim(net_data)
