+++ R/download_forcis_db.R
@@ -55,9 +55,6 @@
 #'
 #' @examples
 #' \dontrun{
-#' # Attach the package ----
-#' library("forcis")
-#'
 #' # Folder in which the database will be saved ----
 #' path_to_save_db <- "data"
 #'
