+++ R/get_version_metadata.R
@@ -15,11 +15,10 @@
 #' @export
 #'
 #' @examples
-#' # Attach the package ----
-#' library("forcis")
-#'
+#' \dontrun{
 #' # Get information for the latest version of the FORCIS database ----
 #' get_version_metadata()
+#' }
 
 get_version_metadata <- function(version = NULL) {
   ## Check arguments ----
