### 2025-04-24 5ffb75c
doc: use read.csv() to import example dataset
