+++ R/select_forcis_columns.R
@@ -21,7 +21,7 @@
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 #'                          package = "forcis")
 #'
-#' net_data <- read.table(file_name, dec = ".", sep = ";")
+#' net_data <- read.csv(file_name)
 #'
 #' # Dimensions of the data.frame ----
 #' dim(net_data)
