+++ R/filter_by_species.R
@@ -17,7 +17,7 @@
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 #'                          package = "forcis")
 #'
-#' net_data <- read.table(file_name, dec = ".", sep = ";")
+#' net_data <- read.csv(file_name)
 #'
 #' # Select a taxonomy ----
 #' net_data <- select_taxonomy(net_data, taxonomy = "VT")
