+++ man/ggmap_data.Rd
@@ -24,7 +24,7 @@ Maps the spatial distribution of FORCIS data.
 file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
 
-net_data <- read.table(file_name, dec = ".", sep = ";")
+net_data <- read.csv(file_name)
 
 # Map data (default) ----
 ggmap_data(net_data)
