+++ man/select_forcis_columns.Rd
@@ -27,7 +27,7 @@ handle it and to speed up some computations.
 file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
 
-net_data <- read.table(file_name, dec = ".", sep = ";")
+net_data <- read.csv(file_name)
 
 # Dimensions of the data.frame ----
 dim(net_data)
