### 2025-04-24 dbcf802
doc: update import of example dataset in vignettes
