+++ vignettes/data-visualization.Rmd
@@ -48,7 +48,7 @@ The vignette will use the plankton net data of the FORCIS database. Let's import
 ```{r 'load-data', echo=FALSE}
 file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
                          package = "forcis")
-net_data <- read.table(file_name, dec = ".", sep = ";")
+net_data <- read.csv(file_name)
 ```
 
 ```{r 'load-data-user', eval=FALSE}
