### 2025-04-24 6155694
Improve example (#103)
* doc: replace example dataset - fix #100

* doc: use read.csv() to import example dataset

* test: update import of example dataset

* doc: update import of example dataset in vignettes
