+++ R/plot_record_by_depth.R
@@ -14,7 +14,7 @@
 #' file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
 #'                          package = "forcis")
 #'
-#' net_data <- read.table(file_name, dec = ".", sep = ";")
+#' net_data <- read.csv(file_name)
 #'
 #' # Plot data by year (example dataset) ----
 #' plot_record_by_depth(net_data)
