+++ tests/testthat/test-compute_frequencies.R
@@ -1,12 +1,10 @@
 ## Data for test ----
 
-df <- read.table(
+df <- read.csv(
   system.file(
     file.path("extdata", "FORCIS_net_sample.csv"),
     package = "forcis"
-  ),
-  dec = ".",
-  sep = ";"
+  )
 )
 
 df <- add_data_type(df, "Net")
@@ -33,12 +31,12 @@ test_that("Test compute_frequencies() for success", {
   expect_message(res <- compute_frequencies(df))
 
   expect_true(is.data.frame(res))
-  expect_equal(ncol(res), 15L)
+  expect_equal(ncol(res), 18L)
   expect_equal(nrow(res), 46324L)
 
   expect_message(res <- compute_frequencies(df, aggregate = FALSE))
 
   expect_true(is.data.frame(res))
-  expect_equal(ncol(res), 18L)
+  expect_equal(ncol(res), 21L)
   expect_equal(nrow(res), 135352L)
 })
