+++ man/plot_record_by_season.Rd
@@ -20,7 +20,7 @@ This function produces a barplot of FORCIS sample records by season.
 file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
 
-net_data <- read.table(file_name, dec = ".", sep = ";")
+net_data <- read.csv(file_name)
 
 # Plot data by year (example dataset) ----
 plot_record_by_season(net_data)
