+++ man/select_taxonomy.Rd
@@ -26,7 +26,7 @@ further information.
 file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"),
                          package = "forcis")
 
-net_data <- read.table(file_name, dec = ".", sep = ";")
+net_data <- read.csv(file_name)
 
 # Dimensions of the data.frame ----
 dim(net_data)
