+++ DESCRIPTION
@@ -47,7 +47,7 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 VignetteBuilder: knitr
 Depends: 
-    R (>= 2.10)
+    R (>= 4.1.0)
 Imports: 
     ggplot2,
     httr2,
