### 2025-04-24 1136cfd
Update CITATION.cff
