### 2025-04-24 3f17456
doc: add main figure
