### 2025-04-24 c9c85fd
doc: add visualization panel
