### 2025-04-24 257016c
rOpenSci review: Add visualizations in the README (#106)
* doc: update workflow figure

* doc: add visualization panel
