+++ README.Rmd
@@ -38,8 +38,7 @@ forcis <img src="man/figures/logo.png" height="120px" align="right" style="float
 <p align="left">
   • <a href="#overview">Overview</a><br>
   • <a href="#installation">Installation</a><br>
-  • <a href="#get-started">Get started</a><br>
-  • <a href="#long-form-documentations">Long-form documentations</a><br>
+  • <a href="#documentation">Documentation</a><br>
   • <a href="#citation">Citation</a><br>
   • <a href="#contributing">Contributing</a><br>
   • <a href="#acknowledgments">Acknowledgments</a><br>
@@ -52,12 +51,14 @@ forcis <img src="man/figures/logo.png" height="120px" align="right" style="float
 
 The goal of the R package `forcis` is to provide an interface to the [FORCIS database](https://zenodo.org/doi/10.5281/zenodo.7390791) on global foraminifera distribution (Chaabane _et al._ 2023). This database includes data on living planktonic foraminifera diversity and distribution in the global oceans from 1910 until 2018 collected using plankton tows, continuous plankton recorder, sediment traps and plankton pump from the global ocean.
 
-This package has been developed for researchers interested in working with the FORCIS database, even without advanced R skills. It provides basic functions to facilitate the handling of this large database, including functions to download, select, filter, homogenize, and visualize the data (Figure 1). It also enables users to explore the spatial distribution and temporal evolution of planktonic foraminifera.
+![Figure 1: Recommended workflow and main features of `forcis`](inst/workflow.png)
 
-![Figure 1. Recommended workflow and main features of `forcis`](inst/workflow.png)
+<br/>
 
-The [Functions](https://frbcesab.github.io/forcis/reference/index.html) page provides a full description of the features available in the package.
+This package has been developed for researchers interested in working with the FORCIS database, even without advanced R skills. It provides basic functions to facilitate the handling of this large database, including functions to download, select, filter, homogenize, and visualize the data. It also enables users to explore the spatial distribution and temporal evolution of planktonic foraminifera.
 
+
+![Figure 2: Overview of visualizations available in `forcis`](inst/visualization.png)
 
 ## Installation
 
@@ -83,13 +84,7 @@ library("forcis")
 ```
 
 
-## Get started
-
-For an overview of the main features of `forcis`, please read the [Get started](https://frbcesab.github.io/forcis/articles/forcis.html) vignette.
-
-
-
-## Long-form documentations
+## Documentation
 
 `forcis` provides five vignettes to learn more about the package:
 
