### 2025-04-24 798ab1c
doc: add use cases in get started
