+++ vignettes/forcis.Rmd
@@ -10,9 +10,9 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  eval     = FALSE,
-  echo     = TRUE,
-  comment  = "#>"
+  eval = TRUE,
+  echo = TRUE,
+  comment = "#>"
 )
 ```
 
@@ -21,7 +21,7 @@ The R package `forcis` is an interface to the [FORCIS database](https://zenodo.o
 
 This package has been developed for researchers interested in working with the FORCIS database, even without advanced R skills. It provides basic functions to facilitate the handling of this large database, including functions to download, select, filter, homogenize, and visualize the data. It also enables users to explore the spatial distribution and temporal evolution of planktonic foraminifera.
 
-This vignette is an overview of the main features.
+This vignette is an overview of the main features of the package.
 
 
 
@@ -41,11 +41,7 @@ remotes::install_github("FRBCesab/forcis")
 
 > The `forcis` package depends on the [`sf`](https://r-spatial.github.io/sf/) package which requires some spatial system libraries (GDAL and PROJ). Please read [this page](https://github.com/r-spatial/sf?tab=readme-ov-file#installing) if you have any trouble to install `forcis`.
 
-
-
-## Setup
-
-First, let's attach the required packages.
+Now let's attach the required packages.
 
 ```{r setup}
 library(forcis)
@@ -60,14 +56,17 @@ The FORCIS database consists of a collection of five `csv` files hosted on [Zeno
 Let's download the latest version of the FORCIS database with `download_forcis_db()`:
 
 ```{r 'download-db', eval=FALSE}
-# Create a data/ folder ----
+# Create a data/ folder in the current directory ----
 dir.create("data")
 
 # Download latest version of the database ----
-download_forcis_db(path = "data", version = NULL)
+download_forcis_db(
+  path = "data",
+  version = NULL
+)
 ```
 
-By default (`version = NULL`), this function downloads the latest version of the database. The database is saved in `data/forcis-db/version-99/`, where `99` is the version number.
+By default (i.e. `version = NULL`), this function downloads the latest version of the database. The database is saved in `data/forcis-db/version-99/`, where `99` is the version number.
 
 **N.B.** The package `forcis` is designed to handle the versioning of the database on Zenodo. Read the [Database versions](https://frbcesab.github.io/forcis/articles/database-versions.html) for more information.
 
@@ -80,9 +79,13 @@ By default (`version = NULL`), this function downloads the latest version of the
 In this vignette, we will use the plankton nets data of the FORCIS database. Let's import the latest release of the data.
 
 ```{r 'load-data', echo=FALSE}
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
-                         package = "forcis")
-net_data <- read.csv(file_name)
+file_name <- system.file(
+  file.path("extdata", "FORCIS_net_sample.csv"),
+  package = "forcis"
+)
+
+net_data <- read.csv(file_name) |>
+  tibble::as_tibble()
 ```
 
 ```{r 'load-data-user', eval=FALSE}
@@ -90,29 +93,167 @@ net_data <- read.csv(file_name)
 net_data <- read_plankton_nets_data(path = "data")
 ```
 
-**N.B.** Here we use a subset of the plankton nets data, not the whole dataset.
+```{r 'print-data'}
+# Print data ----
+net_data
+```
+
+**N.B.** For this vignette, we use a subset of the plankton nets data, not the whole dataset.
 
 
 
-## Select a taxonomy
[…]
