### 2025-04-24 c7389f6
Update CITATION.cff
