### 2025-04-24 caaac96
doc: improve vignette
