+++ vignettes/select-and-filter-data.Rmd
@@ -10,16 +10,16 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  eval     = TRUE,
-  echo     = TRUE,
-  comment  = "#>",
-  dpi       = 120,
+  eval = TRUE,
+  echo = TRUE,
+  comment = "#>",
+  dpi = 120,
   fig.align = "center",
   out.width = "80%"
 )
 ```
 
-The package `forcis` provides [a lot of functions](https://frbcesab.github.io/forcis/reference/index.html#select-and-filters-tools) to filter, reshape, and select FORCIS data. This vignette shows how to use these functions.
+The package `forcis` provides [a lot of functions](https://frbcesab.github.io/forcis/reference/index.html#select-and-filters-tools) to filter, reshape, and select FORCIS data. This vignette shows how to use these functions. With the exception of `select_taxonomy()`, all functions presented in this vignette are optional and depend on your research questions. You can filter data by species, time range, ocean, etc.
 
 
 ## Setup
@@ -43,8 +43,10 @@ download_forcis_db(path = "data", version = NULL)
 The vignette will use the plankton nets data of the FORCIS database. Let's import the latest release of the data.
 
 ```{r 'load-data', echo=FALSE}
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
-                         package = "forcis")
+file_name <- system.file(
+  file.path("extdata", "FORCIS_net_sample.csv"),
+  package = "forcis"
+)
 net_data <- read.csv(file_name)
 ```
 
@@ -60,33 +62,49 @@ net_data <- read_plankton_nets_data(path = "data")
 
 ### Select a taxonomy
 
-The FORCIS database provides three different taxonomies: `LT` (lumped taxonomy), `VT` (validated taxonomy) and `OT` (original taxonomy). See the [associated data paper](https://doi.org/10.1038/s41597-023-02264-2) for further information.
+The FORCIS database provides three different taxonomies: 
 
-After importing the data and before going any further, the next step involves choosing the taxonomic level for the analyses. Let's use the function `select_taxonomy()` to select the **VT** taxonomy (validated taxonomy):
+- `OT`: original taxonomy, i.e. the initial list of species names and attributes (e.g., shell pigmentation, coiling direction) as reported in various datasets and studies.
+- `VT`: validated taxonomy, i.e. a refined version of the original taxonomy that resolves issues of synonymy (different names for the same taxon) and shifting taxonomic concepts.
+- `LT`: lumped taxonomy, i.e. a simplified version of the validated taxonomy. It merges taxa that are difficult to distinguish across datasets (morphospecies), ensuring consistency and comparability in broader analyses. 
+
+See the [associated data paper](https://doi.org/10.1038/s41597-023-02264-2) for further information.
+
+After importing the data and before going any further, the next step involves choosing the taxonomic level for the analyses. **This is mandatory to avoid duplicated records**.
+
+Let's use the function `select_taxonomy()` to select the **VT** taxonomy (validated taxonomy):
 
 ```{r 'select-taxo'}
 # Select taxonomy ----
-net_data <- select_taxonomy(net_data, taxonomy = "VT")
+net_data_vt <- net_data |>
+  select_taxonomy(taxonomy = "VT")
+
+net_data_vt
 ```
 
 
 
 ### Select required columns
 
[…]
