### 2025-04-24 FRBCesab/forcis#94 @ahasverus
Done in #107 and #108
