### 2025-04-24 b3bb98c
doc: change figure paths - fix #109
