### 2025-04-24 bee0f82
rOpenSci review: Add reviewer 2 in DESCRIPTION (#111)
