### 2025-04-24 62005f5
style: use air to format vignettes
