+++ vignettes/data-visualization.Rmd
@@ -9,15 +9,15 @@ vignette: >
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
-  collapse   = TRUE,
-  eval       = TRUE,
-  echo       = TRUE,
-  comment    = "#>",
-  dpi        = 150,
-  fig.align  = "center",
-  out.width  = "80%",
+  collapse = TRUE,
+  eval = TRUE,
+  echo = TRUE,
+  comment = "#>",
+  dpi = 150,
+  fig.align = "center",
+  out.width = "80%",
   fig.height = 4,
-  fig.width  = 7
+  fig.width = 7
 )
 ```
 
@@ -46,8 +46,10 @@ download_forcis_db(path = "data", version = NULL)
 The vignette will use the plankton net data of the FORCIS database. Let's import the latest release of the data.
 
 ```{r 'load-data', echo=FALSE}
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
-                         package = "forcis")
+file_name <- system.file(
+  file.path("extdata", "FORCIS_net_sample.csv"),
+  package = "forcis"
+)
 net_data <- read.csv(file_name)
 ```
 
@@ -121,14 +123,16 @@ indian_net_data_map <- net_data %>%
   ggtitle("FORCIS net data - Indian Ocean")
 
 # Save as PNG ----
-ggsave(indian_net_data_map, 
-       filename = "indian_net_data_map.png",
-       width    = 20, 
-       height   = 11, 
-       units    = "cm", 
-       dpi      = 300,
-       scale    = 1.5,
-       bg       = "white")
+ggsave(
+  indian_net_data_map,
+  filename = "indian_net_data_map.png",
+  width = 20,
+  height = 11,
+  units = "cm",
+  dpi = 300,
+  scale = 1.5,
+  bg = "white"
+)
 ```
 
 
