### 2025-04-24 3f0de74
doc: homogeneize object names in all vignettes
