### 2025-04-24 0c30ede
rOpenSci review: Final review of documentation (#112)
* style: use air to format vignettes

* doc: homogeneize object names in all vignettes
