+++ vignettes/sampling-devices.Rmd
@@ -10,9 +10,9 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  eval     = FALSE,
-  echo     = TRUE,
-  comment  = "#>"
+  eval = FALSE,
+  echo = TRUE,
+  comment = "#>"
 )
 ```
 
