+++ vignettes/data-visualization.Rmd
@@ -9,15 +9,15 @@ vignette: >
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
-  collapse   = TRUE,
-  eval       = TRUE,
-  echo       = TRUE,
-  comment    = "#>",
-  dpi        = 150,
-  fig.align  = "center",
-  out.width  = "80%",
+  collapse = TRUE,
+  eval = TRUE,
+  echo = TRUE,
+  comment = "#>",
+  dpi = 150,
+  fig.align = "center",
+  out.width = "80%",
   fig.height = 4,
-  fig.width  = 7
+  fig.width = 7
 )
 ```
 
@@ -46,8 +46,10 @@ download_forcis_db(path = "data", version = NULL)
 The vignette will use the plankton net data of the FORCIS database. Let's import the latest release of the data.
 
 ```{r 'load-data', echo=FALSE}
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
-                         package = "forcis")
+file_name <- system.file(
+  file.path("extdata", "FORCIS_net_sample.csv"),
+  package = "forcis"
+)
 net_data <- read.csv(file_name)
 ```
 
@@ -95,18 +97,18 @@ This function works with the output of various functions available in the `forci
 
 ```{r 'ggmap-filtered-data', echo=TRUE, eval=FALSE}
 # Filter net data ----
-indian_net_data <- filter_by_ocean(net_data, ocean = "Indian Ocean")
+net_data_indian <- filter_by_ocean(net_data, ocean = "Indian Ocean")
 
 # Map filtered data ----
-ggmap_data(indian_net_data)
+ggmap_data(net_data_indian)
 ```
 
 Note that the `forcis` package is pipe-friendly.
 
 ```{r 'ggmap-filtered-data-pipe', eval=FALSE}
 # Same as before, but w/ the pipe ----
-net_data %>%
-  filter_by_ocean(ocean = "Indian Ocean") %>%
+net_data |>
+  filter_by_ocean(ocean = "Indian Ocean") |>
   ggmap_data()
 ```
 
@@ -115,20 +117,22 @@ You can export this map with the function `ggsave()` of the package `ggplot2`.
 
 ```{r 'ggmap-save', eval=FALSE}
 # Map filtered data ----
-indian_net_data_map <- net_data %>%
-  filter_by_ocean(ocean = "Indian Ocean") %>%
+net_data_indian_map <- net_data |>
+  filter_by_ocean(ocean = "Indian Ocean") |>
   ggmap_data() +
   ggtitle("FORCIS net data - Indian Ocean")
 
 # Save as PNG ----
-ggsave(indian_net_data_map, 
-       filename = "indian_net_data_map.png",
-       width    = 20, 
-       height   = 11, 
-       units    = "cm", 
-       dpi      = 300,
-       scale    = 1.5,
-       bg       = "white")
+ggsave(
+  net_data_indian_map,
+  filename = "net_data_indian_map.png",
+  width = 20,
+  height = 11,
+  units = "cm",
+  dpi = 300,
+  scale = 1.5,
+  bg = "white"
+)
 ```
 
 
