+++ vignettes/data-conversion.Rmd
@@ -10,9 +10,9 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  eval     = TRUE,
-  echo     = TRUE,
-  comment  = "#>"
+  eval = TRUE,
+  echo = TRUE,
+  comment = "#>"
 )
 ```
 
@@ -57,8 +57,10 @@ download_forcis_db(path = "data", version = NULL)
 The vignette will use the plankton net data of the FORCIS database. Let's import the latest release of the data.
 
 ```{r 'load-data', echo=FALSE}
-file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
-                         package = "forcis")
+file_name <- system.file(
+  file.path("extdata", "FORCIS_net_sample.csv"),
+  package = "forcis"
+)
 net_data <- read.csv(file_name)
 ```
 
@@ -76,7 +78,8 @@ Let's use the function `select_taxonomy()` to select the **VT** taxonomy (valida
 
 ```{r 'select-taxo'}
 # Select taxonomy ----
-net_data <- select_taxonomy(net_data, taxonomy = "VT")
+net_data_vt <- net_data |>
+  select_taxonomy(taxonomy = "VT")
 ```
 
 Once the data contain counts from the same taxonomic level, we can proceed with the conversion functions: `compute_*()`.
@@ -90,17 +93,19 @@ This function converts all counts into raw abundances, using sampling metadata s
 
 ```{r 'compute-abundance'}
 # Convert species counts in raw abundance ----
-net_data_raw_ab <- compute_abundances(net_data, aggregate = TRUE)
+net_data_vt_raw_ab <- net_data_vt |>
+  compute_abundances(aggregate = TRUE)
 ```
 
 ```{r 'exploration'}
 # Format ----
-dim(net_data)
-dim(net_data_raw_ab)
+dim(net_data_vt)
+dim(net_data_vt_raw_ab)
 
 # Header ----
-net_data_raw_ab <- as.data.frame(net_data_raw_ab)
-head(net_data_raw_ab)
+net_data_vt_raw_ab |>
+  as.data.frame() |>
+  head()
 ```
 
 The functions `compute_*()` output a table in a long-format as well as a message reporting the amount of data that could not be converted because of missing metadata.
@@ -113,8 +118,8 @@ This function transforms all counts into number concentration abundances. It als
 
 ```{r 'compute-concetration'}
 # Convert species counts in number concentration ----
-net_data_n_conc <- compute_concentrations(net_data, aggregate = TRUE)
-
+net_data_vt_n_conc <- net_data_vt |>
+  compute_concentrations(aggregate = TRUE)
 ```
 
 
@@ -124,5 +129,6 @@ This function computes relative abundance for each species using total assemblag
 
 ```{r 'compute-frequency'}
 # Convert species counts in relative abundance ----
-net_data_rel_ab <- compute_frequencies(net_data, aggregate = TRUE)
+net_data_rel_ab <- net_data_vt |>
+  compute_frequencies(aggregate = TRUE)
 ```
