### 2025-04-24 addd4a6
Refactor(tests): implement offline test mocking (#91) (#101)
* chore(dev): add auto format for vscode/positron

* tests: refactor mocking data for offline mode

* chore: fix checks
