+++ tests/testthat/test-read_plankton_nets_data.R
@@ -8,25 +8,30 @@ df <- data.frame("crassula" = rep(1, 2), "dehiscens" = rep(1, 2))
 test_that("Test read_plankton_nets_data() for error", {
   create_tempdir()
 
-  expect_error(
-    read_plankton_nets_data(version = "07"),
-    paste0(
-      "The directory './forcis-db/version-07' does not exist. ",
-      "Please check the argument 'path' or use the function ",
-      "'download_forcis_db()'."
-    ),
-    fixed = TRUE
-  )
+  with_mocked_bindings(
+    get_metadata = get_metadata_mock,
+    {
+      expect_error(
+        read_plankton_nets_data(version = "07"),
+        paste0(
+          "The directory './forcis-db/version-07' does not exist. ",
+          "Please check the argument 'path' or use the function ",
+          "'download_forcis_db()'."
+        ),
+        fixed = TRUE
+      )
 
-  dir.create(file.path("forcis-db", "version-07"), recursive = TRUE)
+      dir.create(file.path("forcis-db", "version-07"), recursive = TRUE)
 
-  expect_error(
-    read_plankton_nets_data(version = "07"),
-    paste0(
-      "The Plankton net dataset does not exist. Please use ",
-      "the function 'download_forcis_db()'."
-    ),
-    fixed = TRUE
+      expect_error(
+        read_plankton_nets_data(version = "07"),
+        paste0(
+          "The Plankton net dataset does not exist. Please use ",
+          "the function 'download_forcis_db()'."
+        ),
+        fixed = TRUE
+      )
+    }
   )
 })
 
@@ -41,24 +46,29 @@ test_that("Test read_plankton_nets_data() for success", {
     row.names = FALSE
   )
 
-  x <- read_plankton_nets_data(version = "07", check_for_update = FALSE)
-
-  expect_true("data.frame" %in% class(x))
-  expect_equal(ncol(x), 3L)
-  expect_equal(nrow(x), 2L)
-
-  expect_true("data_type" %in% colnames(x))
-  expect_true("Net" %in% x$"data_type")
-
-  expect_message(
+  with_mocked_bindings(
+    get_metadata = get_metadata_mock,
     {
-      x <- read_plankton_nets_data(version = "07")
-    },
-    paste0(
-      "A newer version of the FORCIS database is available. ",
-      "Use 'download_forcis_db(version = NULL)' to download ",
-      "it."
-    ),
-    fixed = TRUE
+      x <- read_plankton_nets_data(version = "07", check_for_update = FALSE)
+
+      expect_true("data.frame" %in% class(x))
+      expect_equal(ncol(x), 3L)
+      expect_equal(nrow(x), 2L)
+
+      expect_true("data_type" %in% colnames(x))
+      expect_true("Net" %in% x$"data_type")
+
+      expect_message(
+        {
+          x <- read_plankton_nets_data(version = "07")
+        },
+        paste0(
+          "A newer version of the FORCIS database is available. ",
+          "Use 'download_forcis_db(version = NULL)' to download ",
+          "it."
+        ),
+        fixed = TRUE
+      )
+    }
   )
 })
