### 2025-04-24 a8edba8
test: suppress message in compute_*()
