### 2025-04-24 57c3c08
rOpenSci: Suppress messages in compute_*() (#118)
