### 2025-04-24 82cd09e
doc: update CHANGELOG
