### 2025-04-24 #660 @ahasverus
Dear @heyairf

Thanks for your constructive comments!

Here are my answers:

**Documentation**

> README includes an overview of the what it does but does not indicate why it was made or who it is for. A sentence or two on how this data could be used might be a nice addition to overview section, though once the target audience is identified this might not be necessary.
>
> I think a small example of the package functionality on the README page or the Get Started page would be a great addition. I had to click around a lot before I understood what the package use would look like, even after running the code on the get started page.

I made three changes in the [`README`](https://github.com/FRBCesab/forcis/blob/main/README.md):

- Add a statement of need (one paragraph) explaining the targetted audience of the package - [PR #84](https://github.com/FRBCesab/forcis/pull/84)
- Add a figure showing the recommended workflow of the package - [PR #105](https://github.com/FRBCesab/forcis/pull/105)
- Add a panel of four plots to show an overview of visualizations available in the package - [PR #106](https://github.com/FRBCesab/forcis/pull/106)

**Get Started page**

> I was still pretty unsure what the purpose and final result of using this package was at the end of this page. I recommend including a small example here! Even as simple as a `head()` and small description of what I am looking at would provide a bit more context to users.

I improved the [Get Started](https://frbcesab.github.io/forcis/articles/forcis.html) vignette by showing function outputs (it's easier now because all functions return `tibble`) and by adding two use cases - [PR #107](https://github.com/FRBCesab/forcis/issues/107)

I hope that with these changes (`README` and this vignette), the purpose of the package is clearer.

> Linking to a short vignette describing the taxonomy codes (similar in scope to the Sampling devices vignette) might be less intimidating than linking to the data paper... or at least providing and indication of which section in the data paper to read about this would be helpful! It was not immediately clear from the paper headings where to find this information.

I added the definition of the three taxonomies - [PR #107](https://github.com/FRBCesab/forcis/issues/107) and [PR #108](https://github.com/FRBCesab/forcis/issues/108)

Is it clearer now?

**Select, reshape, filter**

> In the selecting columns section, it was unclear to me if you were showing two different selection methods, or a multistep selection method... are you showing a filter of VT data and then forcis columns? perhaps naming the objects differently and/or expanding the commented lines would make this clearer.

I added a sentence in this vignette explaining that the use if these functions are optional and depend on the user questions - [PR #108](https://github.com/FRBCesab/forcis/issues/108)

> Also, the same selection code was used on the Get Started page but the object was given a different name... could consider keeping this consistent.

Thanks. Done in [PR #107](https://github.com/FRBCesab/forcis/issues/107), [PR #108](https://github.com/FRBCesab/forcis/issues/108), and [PR #112](https://github.com/FRBCesab/forcis/issues/112)

**Conversion / visualization**

> These vignettes were the first time I felt I understood the point of the package/some potential uses. Having a small visualization example like these on the README or Get Started to show a start to finish workflow to a sensible end result would be helpful even if its redundant with the more detailed vignettes.

I hope that changes made on the `README` and the Get started vignette help user to better understand the prupose of the package.

**Function documentation and examples**

> Looks great! Some of your examples include library("forcis") before reading example data and some don't...perhaps commit one way or the other. I think it's usually pretty standard to exclude these for the package itself, but should be done you're attaching a dependency (like for ggplot2 in this example)

Thanks! Done in [PR #99](https://github.com/FRBCesab/forcis/pull/99)

> Quite a lot of `\dontrun{}`'s in the examples .. which could potentially be an issue if planning to submit to CRAN. Would any of the suggestions in this chapter be appropriate to reduce these? the http and API complexities in your package examples/tests are far beyond the scope of anything I've ever written, so please disregard this feedback if it is irrelevant.

I removed the `\dontrun{}` in some examples (see [PR #99](https://github.com/FRBCesab/forcis/pull/99)) but for functions querying the Zenodo API, I need to use `\dontrun{}`.

**Tests**

> I see you have added a skip() for the test that was causing failure in the earlier review, I did not get any failures on my end because of this! Though I get the same error when I run the test without it.

All unit tests passed now! [PR #101](https://github.com/FRBCesab/forcis/pull/101)

> All other tests pass; however, when running devtools::test() I get the message "Relative counts from 24 samples could not be converted because of missing data on total assemblage" quite a few times. Is it worth suppressing these to reduce clutter?

Thanks! Done in [PR #118](https://github.com/FRBCesab/forcis/pull/118)

And thanks again for your time!
Nicolas
