### 2025-04-25 #660 @ahasverus
In addition to my previous reponses, here is the updated [NEWS.md](https://github.com/FRBCesab/forcis/blob/main/NEWS.md)
