### 2025-04-25 ac7264e
ci: add real tests
