+++ .github/workflows/R-CMD-check-schedule.yaml
@@ -1,0 +1,53 @@
+# Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
+# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
+on:
+    push:
+      branches: [main, master]
+    schedule:
+      - cron: "15 0 * * *"
+  
+name: R-CMD-check
+  
+jobs:
+    R-CMD-check:
+      runs-on: ${{ matrix.config.os }}
+  
+      name: ${{ matrix.config.os }} (${{ matrix.config.r }})
+  
+      strategy:
+        fail-fast: false
+        matrix:
+          config:
+            - {os: macOS-latest,   r: 'release'}
+            - {os: windows-latest, r: 'release'}
+            - {os: ubuntu-latest,  r: 'devel', http-user-agent: 'release'}
+            - {os: ubuntu-latest,  r: 'release'}
+            - {os: ubuntu-latest,  r: 'oldrel-1'}
+  
+      env:
+        GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+        R_KEEP_PKG_SOURCE: yes
+  
+      steps:
+        - uses: actions/checkout@v4
+  
+        - uses: r-lib/actions/setup-pandoc@v2
+  
+        - uses: r-lib/actions/setup-r@v2
+          with:
+            r-version: ${{ matrix.config.r }}
+            http-user-agent: ${{ matrix.config.http-user-agent }}
+            use-public-rspm: true
+  
+        - uses: r-lib/actions/setup-r-dependencies@v2
+          with:
+            extra-packages: any::rcmdcheck
+            needs: check
+        
+        - name: Folder with real tests
+          run: |
+            rm -rf tests/testthat && cp -r real-tests tests/testthat
+  
+        - uses: r-lib/actions/check-r-package@v2
+          with:
+            upload-snapshots: true
