### 2025-04-25 40f94bb
chore(ci): make it run on PR (tmp)
