### 2025-04-25 68e106d
ci: add real tests (#122)
* ci: add real tests

* chore(ci): make it run on PR (tmp)

* chore(ci): use one GH runner (Real Tests)
