### 2025-04-25 1bd8c45
ci: enable real tests on pr
