+++ .github/workflows/R-CMD-check-schedule.yaml
@@ -3,8 +3,8 @@
 on:
     push:
       branches: [main, master]
-    # pull_request: # TO BE REMOVED. Testing on PR
-    #   branches: [main, master] 
+    pull_request:
+      branches: [main, master] 
     schedule:
       - cron: "15 0 * * *"
   
