### 2025-04-25 #660 @ahasverus
> Hi @ahasverus ,
>
> Your new README is fantastic! The flowchart is great at showing how or why a user could use the package. The changes made to the rest of the documentation also look great and have added more clarity. The package is now far more self-contained. Well done!

Thank you so much!
