### 2025-05-04 #660 @beatrizmilz
Thank you, @ahasverus, for addressing the reviewers’ comments!

---

@sckott and @heyairf, thank you both again for your reviews and for responding after the recent changes.
Could you please indicate if you approve the package using the template?
 https://devguide.ropensci.org/approval2template.html
Thank you so much!
