### 2025-05-04 #660 @maurolepore
Hey all, I'm wrapping up my EiC rotation today and before I hand it over to the next EiC I wanted to thank you all for your work.
