### 2025-05-13 #660 @beatrizmilz
> Hey all, I'm wrapping up my EiC rotation today and before I hand it over to the next EiC I wanted to thank you all for your work.

Thank you Mauro!! @maurolepore

Also, thank you both @sckott  and @heyairf for the review!
