### 2025-05-13 #660 @ropensci-review-bot
Approved! Thanks @ahasverus for submitting and @sckott, @heyairf for your reviews! :grin:
