### 2025-05-14 #660 @ahasverus
I have finalized the transfer of the repo to the GH organization [rOpenSci](https://github.com/ropensci/forcis).
I just want to thank you all for this review (especially @sckott and @heyairf).
Nicolas
